Refactor App.tsx into layout and contexts
Extracted massive state and logic from App.tsx into PreferencesContext, UIContext, UpdateContext, AppLayout, and MainContent.
This improves code organization, maintainability, and readability.
Preserved all existing functionality and addressed type errors.
