+++ renderer/src/App.tsx
@@ -1,3027 +1,18 @@
-import { useState, useEffect, useMemo, useRef, useCallback } from 'react';
-import { useGameLibrary } from './hooks/useGameLibrary';
-import { useFullscreen } from './hooks/useFullscreen';
-import { LibraryGrid } from './components/LibraryGrid';
-import { LibraryListView } from './components/LibraryListView';
-import { LibraryCarousel } from './components/LibraryCarousel';
-import { LibraryCoverFlow } from './components/LibraryCoverFlow';
-import { RightClickMenu } from './components/RightClickMenu';
-import { GameContextMenu } from './components/GameContextMenu';
-import { AddGameModal } from './components/AddGameModal';
-import { GameDetailsPanel } from './components/GameDetailsPanel';
-import { LINK_DISPLAY_ORDER, DEFAULT_VISIBLE_LINK_TYPES } from './components/GameLinks';
-import { GameMetadataEditor } from './components/GameMetadataEditor';
-import { SteamConfigModal } from './components/SteamConfigModal';
-import { CategoriesEditor } from './components/CategoriesEditor';
-import { TopBar } from './components/TopBar';
-import { MenuBar } from './components/MenuBar';
-import { TopBarPositions } from './components/TopBarContextMenu';
-import { UpdateLibraryModal } from './components/UpdateLibraryModal';
-import { OnyxSettingsModal } from './components/OnyxSettingsModal';
-import { APISettingsModal } from './components/APISettingsModal';
-import { MetadataSearchModal } from './components/MetadataSearchModal';
-import { ImportWorkbenchV2 as ImportWorkbench } from './components/importer/ImportWorkbenchV2';
-import { GameManager } from './components/GameManager';
-import { ConfirmationDialog } from './components/ConfirmationDialog';
-import { BugReportModal } from './components/BugReportModal';
-import { WelcomeScreen } from './components/WelcomeScreen';
-import { MissingGamesModal } from './components/MissingGamesModal';
-import { UpdateNotificationModal } from './components/UpdateNotificationModal';
-import { LibraryTutorialModal } from './components/LibraryTutorialModal';
-import { Game, ExecutableFile, GameMetadata } from './types/game';
-import { areAPIsConfigured } from './utils/apiValidation';
+import { PreferencesProvider } from './contexts/PreferencesContext';
+import { UIProvider } from './contexts/UIContext';
+import { UpdateProvider } from './contexts/UpdateContext';
+import { AppLayout } from './components/layout/AppLayout';
 
 function App() {
-  // Main App Component
-  const { games, loading, error, reorderGames, addCustomGame, loadLibrary, deleteGame, updateGameInState } = useGameLibrary();
-
-  // Initialize fullscreen and gamepad support
-  useFullscreen();
-
-  const [activeGameId, setActiveGameId] = useState<string | null>(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [launchingGameId, setLaunchingGameId] = useState<string | null>(null);
-  const [runningGames, setRunningGames] = useState<Set<string>>(new Set());
-
-  // Scanning state
-  const [, setIsScanningSteam] = useState(false);
-
-  // Folder scan state (for ImportWorkbench)
-  const [, setImportWorkbenchFolderPath] = useState<string | undefined>(undefined);
-
-  // Metadata editor state
-  const [isMetadataEditorOpen, setIsMetadataEditorOpen] = useState(false);
-  const [selectedExecutable, setSelectedExecutable] = useState<ExecutableFile | null>(null);
-
-  // Steam config modal state
-  const [isSteamConfigOpen, setIsSteamConfigOpen] = useState(false);
-
-  const [scannedSteamGames, setScannedSteamGames] = useState<Array<any>>([]);
-  const [, setImportAppType] = useState<'steam' | 'xbox' | 'other'>('steam');
-
-
-  // Categories editor state
-  const [isCategoriesEditorOpen, setIsCategoriesEditorOpen] = useState(false);
-  const [editingCategoriesGame, setEditingCategoriesGame] = useState<Game | null>(null);
-
-  // Metadata search modal state
-  const [isMetadataSearchOpen, setIsMetadataSearchOpen] = useState(false);
-  const [fixingGame, setFixingGame] = useState<Game | null>(null);
-
-  // Toast notification state
-  const [toast, setToast] = useState<{ message: string; type: 'success' | 'error' } | null>(null);
-
-  // Update notification state
-  const [updateNotification, setUpdateNotification] = useState<{
-    version: string;
-    status: 'available' | 'downloading' | 'downloaded' | 'error';
-    error?: string;
-  } | null>(null);
-  const [currentVersion, setCurrentVersion] = useState<string | null>(null);
-  const [changelogSource, setChangelogSource] = useState<string | null>(null);
-  const [changelogLoading, setChangelogLoading] = useState(false);
-  const [changelogError, setChangelogError] = useState<string | null>(null);
-  const [isUpdateModalTest, setIsUpdateModalTest] = useState(false);
-
-  // Search and view state
-  const [searchQuery, setSearchQuery] = useState('');
-  const [viewMode, setViewMode] = useState<'grid' | 'list' | 'logo' | 'carousel' | 'coverflow'>('grid');
-  const [activeSection] = useState('library');
-  const [showTopBar] = useState(false);
-  const [isUpdateLibraryOpen, setIsUpdateLibraryOpen] = useState(false);
-  const [isOnyxSettingsOpen, setIsOnyxSettingsOpen] = useState(false);
-  const [isImportWorkbenchOpen, setIsImportWorkbenchOpen] = useState(false);
-  const [showLibraryTutorial, setShowLibraryTutorial] = useState(false);
-  const [autoStartScan, setAutoStartScan] = useState(false);
-  const [isGameManagerOpen, setIsGameManagerOpen] = useState(false);
-  const [gameManagerInitialGameId, setGameManagerInitialGameId] = useState<string | null>(null);
-  const [gameManagerInitialTab, setGameManagerInitialTab] = useState<'images' | 'metadata' | 'modManager'>('images');
-  const [onyxSettingsInitialTab, setOnyxSettingsInitialTab] = useState<'general' | 'appearance' | 'apis' | 'apps' | 'about'>('general');
-  const [isAPISettingsOpen, setIsAPISettingsOpen] = useState(false);
-  const [isBugReportOpen, setIsBugReportOpen] = useState(false);
-  const [gridSize, setGridSize] = useState(120);
-  const [logoSize, setLogoSize] = useState(100);
-  const [pinnedCategories, setPinnedCategories] = useState<string[]>([]);
-  const [hideVRTitles, setHideVRTitles] = useState(true);
-  const [hideAppsTitles, setHideAppsTitles] = useState(true);
-  const [hideGameTitles, setHideGameTitles] = useState(false);
-  const [gameTilePadding, setGameTilePadding] = useState(3);
-  const [selectedBoxArtSize, setSelectedBoxArtSize] = useState(25);
-  const [showLogoOverBoxart, setShowLogoOverBoxart] = useState(true);
-  const [logoPosition, setLogoPosition] = useState<'top' | 'middle' | 'bottom' | 'underneath'>('middle');
-  const [logoBackgroundColor, setLogoBackgroundColor] = useState('#374151');
-  const [logoBackgroundOpacity, setLogoBackgroundOpacity] = useState(100);
-  const [backgroundBlur, setBackgroundBlur] = useState(40);
-  const [backgroundBrightnessByView, setBackgroundBrightnessByView] = useState<Record<'grid' | 'list' | 'logo' | 'carousel' | 'coverflow', number>>({
-    grid: 0.3,
-    list: 0.3,
-    logo: 0.3,
-    carousel: 0.3,
-    coverflow: 0.3,
-  });
-  const [showCarouselDetails, setShowCarouselDetails] = useState(true);
-  const [showCarouselLogos, setShowCarouselLogos] = useState(true);
-  const [detailsBarSize, setDetailsBarSize] = useState(14);
-  const [carouselLogoSize, setCarouselLogoSize] = useState(100);
-  const [carouselButtonSize, setCarouselButtonSize] = useState(14);
-  const [carouselDescriptionSize, setCarouselDescriptionSize] = useState(18);
-  const [startupProgress, setStartupProgress] = useState<{ message: string } | null>(null);
-  const [gridDescriptionSize] = useState(14);
-  const defaultListViewOptions = {
-    showDescription: true,
-    showCategories: false,
-    showPlaytime: true,
-    showReleaseDate: true,
-    showGenres: true,
-    showPlatform: false,
-    showLauncher: true,
-    showLogos: false,
-    titleTextSize: 18,
-  };
-  // Link management (source of truth from Settings; respected by all views)
-  const [linkDisplayOrder, setLinkDisplayOrder] = useState<string[]>(LINK_DISPLAY_ORDER);
-  const [visibleLinkTypes, setVisibleLinkTypes] = useState<Record<string, boolean>>(DEFAULT_VISIBLE_LINK_TYPES);
-
-  // Right panel (GameDetailsPanel) settings
-  const [rightPanelLogoSize, setRightPanelLogoSize] = useState(100);
-  const [rightPanelBoxartPosition, setRightPanelBoxartPosition] = useState<'left' | 'right' | 'none'>('right');
-  const [rightPanelBoxartSize, setRightPanelBoxartSize] = useState(120);
-  const [rightPanelTextSize, setRightPanelTextSize] = useState(14);
-  const [rightPanelButtonSize, setRightPanelButtonSize] = useState(14);
-  const [rightPanelButtonLocation, setRightPanelButtonLocation] = useState<'left' | 'middle' | 'right'>('right');
-  const [detailsPanelOpacity, setDetailsPanelOpacity] = useState(80);
-  const [isViewFlippedByView, setIsViewFlippedByView] = useState<Record<'grid' | 'list' | 'logo' | 'carousel' | 'coverflow', boolean>>({
-    grid: false,
-    list: false,
-    logo: false,
-    carousel: false,
-    coverflow: false,
-  });
-  // Button colors per view
-  const [rightPanelButtonColors, setRightPanelButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  const [carouselButtonColors, setCarouselButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  const [gridButtonColors, setGridButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  const [listButtonColors, setListButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  const [logoButtonColors, setLogoButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  // Cover Flow (simplified menu)
-  const [coverFlowCoverSize, setCoverFlowCoverSize] = useState(300);
-  const [coverFlowReflection, setCoverFlowReflection] = useState(60);
-  const [coverFlowVerticalOffset, setCoverFlowVerticalOffset] = useState(0);
-  const [coverFlowSideOpacity, setCoverFlowSideOpacity] = useState(100);
-  const [coverFlowShowButtons, setCoverFlowShowButtons] = useState(true);
-  const [coverFlowButtonPosition, setCoverFlowButtonPosition] = useState<'left' | 'middle' | 'right'>('middle');
-  const [coverFlowButtonColors, setCoverFlowButtonColors] = useState<{ playColor?: string; editColor?: string; modManagerColor?: string }>({ playColor: '#0ea5e9', editColor: '#6b7280', modManagerColor: '#a855f7' });
-  // Top bar element positions
-  const [topBarPositions, setTopBarPositions] = useState<TopBarPositions>({
-    searchBar: 'left',
-    sortBy: 'left',
-    launcher: 'left',
-    categories: 'left',
-  });
-  // Game details panel divider settings per view
-  const [fanartHeightByView, setFanartHeightByView] = useState<Record<'grid' | 'list' | 'logo', number>>({
-    grid: 320,
-    list: 320,
-    logo: 320,
-  });
-  const [descriptionWidthByView, setDescriptionWidthByView] = useState<Record<'grid' | 'list' | 'logo', number>>({
-    grid: 50,
-    list: 50,
-    logo: 50,
-  });
-  const [panelWidthByViewState, setPanelWidthByViewState] = useState<Record<'grid' | 'list' | 'logo' | 'carousel' | 'coverflow', number>>({
-    grid: 800,
-    list: 800,
-    logo: 800,
-    carousel: 800,
-    coverflow: 800,
-  });
-  const [showCategoriesByView, setShowCategoriesByView] = useState<Record<string, boolean>>({
-    grid: false,
-    list: false,
-    logo: false,
-  });
-  const [categoriesPositionByView, setCategoriesPositionByView] = useState<Record<string, 'top' | 'bottom'>>({
-    grid: 'top',
-    list: 'top',
-    logo: 'top'
-  });
-  const [categoriesAlignmentByView, setCategoriesAlignmentByView] = useState<Record<string, 'left' | 'center' | 'right'>>({
-    grid: 'left',
-    list: 'left',
-    logo: 'left'
-  });
-  const [categoriesSizeByView, setCategoriesSizeByView] = useState<Record<string, number>>({
-    grid: 12,
-    list: 12,
-    logo: 12
-  });
-
-  // Launch confirmation state
-  const [confirmGameLaunch, setConfirmGameLaunch] = useState(false);
-  const [launchConfirmation, setLaunchConfirmation] = useState<{ game: Game } | null>(null);
-
-  // Missing games state
-  const [missingGames, setMissingGames] = useState<Array<{
-    id: string;
-    title: string;
-    exePath?: string;
-    platform?: string;
-    source?: string;
-  }> | null>(null);
-
-  // Found games state
-  const [foundGames, setFoundGames] = useState<Array<any> | null>(null);
-
-
-
-  // Get current view's divider settings
-  const currentFanartHeight = (viewMode === 'grid' || viewMode === 'list' || viewMode === 'logo') ? fanartHeightByView[viewMode] : 320;
-  const currentDescriptionWidth = (viewMode === 'grid' || viewMode === 'list' || viewMode === 'logo') ? descriptionWidthByView[viewMode] : 50;
-  const currentPanelWidth = (viewMode === 'grid' || viewMode === 'list' || viewMode === 'logo') ? panelWidthByViewState[viewMode] : 800;
-  const currentBackgroundBrightness = backgroundBrightnessByView[viewMode] ?? 0.3;
-
-  // Set background blur to 0 when switching to carousel/coverflow mode and sync divider widths when view changes
-  useEffect(() => {
-    if ((viewMode === 'carousel' || viewMode === 'coverflow') && backgroundBlur !== 0) {
-      setBackgroundBlur(0);
-    }
-    // Sync panelWidth to current view's setting
-    if (viewMode !== 'carousel' && viewMode !== 'coverflow') {
-      const viewSpecificWidth = panelWidthByViewState[viewMode];
-      setPanelWidth(viewSpecificWidth);
-    }
-  }, [viewMode]);
-  const [backgroundMode, setBackgroundMode] = useState<'image' | 'color'>('image');
-  const [backgroundColor, setBackgroundColor] = useState('#000000');
-  const [listViewOptions, setListViewOptions] = useState(defaultListViewOptions);
-  const [listViewSize, setListViewSize] = useState(128);
-  const [panelWidth, setPanelWidth] = useState(800);
-  const [rightClickMenu, setRightClickMenu] = useState<{ x: number; y: number } | null>(null);
-  const [gameContextMenu, setGameContextMenu] = useState<{ x: number; y: number; game: Game } | null>(null);
-  const [isInitialLoad, setIsInitialLoad] = useState(true);
-  const [autoSizeToFit, setAutoSizeToFit] = useState(false);
-  const gridContainerRef = useRef<HTMLDivElement>(null);
-  const currentResolutionRef = useRef<string>(window.screen.height >= 2160 ? '4K' : window.screen.height >= 1440 ? '1440p' : window.screen.height >= 1080 ? '1080p' : '720p');
-  const baselineDefaultsRef = useRef<any>(null);
-
-  const applyPreferences = (prefs: any, options?: { markInitialLoad?: boolean }) => {
-    if (prefs.gridSize) setGridSize(prefs.gridSize);
-    if (prefs.logoSize) setLogoSize(prefs.logoSize);
-    if (prefs.pinnedCategories) setPinnedCategories(prefs.pinnedCategories);
-    if (prefs.hideVRTitles !== undefined) setHideVRTitles(prefs.hideVRTitles);
-    if (prefs.hideAppsTitles !== undefined) setHideAppsTitles(prefs.hideAppsTitles);
-    if (prefs.hideGameTitles !== undefined) setHideGameTitles(prefs.hideGameTitles);
-    if (prefs.gameTilePadding !== undefined) setGameTilePadding(prefs.gameTilePadding);
-    if (prefs.showCategoriesInGameListByView !== undefined) setShowCategoriesByView(prefs.showCategoriesInGameListByView);
-    if (prefs.categoriesPositionByView !== undefined) setCategoriesPositionByView(prefs.categoriesPositionByView);
-    if (prefs.categoriesAlignmentByView !== undefined) setCategoriesAlignmentByView(prefs.categoriesAlignmentByView);
-    if (prefs.categoriesSizeByView !== undefined) setCategoriesSizeByView(prefs.categoriesSizeByView);
-    if (prefs.showLogoOverBoxart !== undefined) setShowLogoOverBoxart(prefs.showLogoOverBoxart);
-    if (prefs.logoPosition !== undefined) setLogoPosition(prefs.logoPosition);
-    if (prefs.logoBackgroundColor !== undefined) setLogoBackgroundColor(prefs.logoBackgroundColor);
-    if (prefs.logoBackgroundOpacity !== undefined) setLogoBackgroundOpacity(prefs.logoBackgroundOpacity);
-    if (prefs.backgroundBlur !== undefined) setBackgroundBlur(prefs.backgroundBlur);
-    if (prefs.backgroundBrightnessByView !== undefined) {
-      setBackgroundBrightnessByView({
-        grid: prefs.backgroundBrightnessByView.grid ?? 0.3,
-        list: prefs.backgroundBrightnessByView.list ?? 0.3,
-        logo: prefs.backgroundBrightnessByView.logo ?? 0.3,
-        carousel: prefs.backgroundBrightnessByView.carousel ?? 0.3,
-        coverflow: prefs.backgroundBrightnessByView.coverflow ?? 0.3,
-      });
-    }
-    if (prefs.showCarouselDetails !== undefined) setShowCarouselDetails(prefs.showCarouselDetails);
-    if (prefs.showCarouselLogos !== undefined) setShowCarouselLogos(prefs.showCarouselLogos);
-    if (prefs.detailsBarSize !== undefined) setDetailsBarSize(prefs.detailsBarSize);
-    if (prefs.carouselLogoSize !== undefined) setCarouselLogoSize(prefs.carouselLogoSize);
-    if (prefs.carouselButtonSize !== undefined) setCarouselButtonSize(prefs.carouselButtonSize);
-    if (prefs.carouselDescriptionSize !== undefined) setCarouselDescriptionSize(prefs.carouselDescriptionSize);
-    // Right panel settings
-    if (prefs.rightPanelLogoSize !== undefined) setRightPanelLogoSize(prefs.rightPanelLogoSize);
-    if (prefs.rightPanelBoxartPosition !== undefined) setRightPanelBoxartPosition(prefs.rightPanelBoxartPosition);
-    if (prefs.rightPanelBoxartSize !== undefined) setRightPanelBoxartSize(prefs.rightPanelBoxartSize);
-    if (prefs.rightPanelTextSize !== undefined) setRightPanelTextSize(prefs.rightPanelTextSize);
-    if (prefs.rightPanelButtonSize !== undefined) setRightPanelButtonSize(prefs.rightPanelButtonSize);
-    if (prefs.rightPanelButtonLocation !== undefined) setRightPanelButtonLocation(prefs.rightPanelButtonLocation);
-    if (prefs.detailsPanelOpacity !== undefined) setDetailsPanelOpacity(prefs.detailsPanelOpacity);
-    if (prefs.rightPanelButtonColors !== undefined) setRightPanelButtonColors(prefs.rightPanelButtonColors);
-    if (prefs.carouselButtonColors !== undefined) setCarouselButtonColors(prefs.carouselButtonColors);
-    if (prefs.gridButtonColors !== undefined) setGridButtonColors(prefs.gridButtonColors);
-    if (prefs.listButtonColors !== undefined) setListButtonColors(prefs.listButtonColors);
-    if (prefs.logoButtonColors !== undefined) setLogoButtonColors(prefs.logoButtonColors);
-    if (prefs.coverFlowCoverSize !== undefined) setCoverFlowCoverSize(prefs.coverFlowCoverSize);
-    if (prefs.coverFlowReflection !== undefined) setCoverFlowReflection(prefs.coverFlowReflection);
-    if (prefs.coverFlowVerticalOffset !== undefined) setCoverFlowVerticalOffset(prefs.coverFlowVerticalOffset);
-    if (prefs.coverFlowSideOpacity !== undefined) setCoverFlowSideOpacity(prefs.coverFlowSideOpacity);
-    if (prefs.coverFlowShowButtons !== undefined) setCoverFlowShowButtons(prefs.coverFlowShowButtons);
-    if (prefs.coverFlowButtonPosition !== undefined) setCoverFlowButtonPosition(prefs.coverFlowButtonPosition);
-    if (prefs.coverFlowButtonColors !== undefined) setCoverFlowButtonColors(prefs.coverFlowButtonColors);
-    if (prefs.isViewFlippedByView !== undefined) {
-      const defaultFlipped = { grid: false, list: false, logo: false, carousel: false, coverflow: false };
-      setIsViewFlippedByView({ ...defaultFlipped, ...prefs.isViewFlippedByView });
-    }
-    // Top bar positions
-    if (prefs.topBarPositions) setTopBarPositions({ ...topBarPositions, ...prefs.topBarPositions });
-    if (prefs.viewMode) setViewMode(prefs.viewMode);
-    if (prefs.backgroundMode) setBackgroundMode(prefs.backgroundMode as 'image' | 'color');
-    if (prefs.backgroundColor) setBackgroundColor(prefs.backgroundColor);
-    if (prefs.listViewOptions) {
-      setListViewOptions({ ...defaultListViewOptions, ...prefs.listViewOptions });
-    } else {
-      setListViewOptions(defaultListViewOptions);
-    }
-    if (prefs.listViewSize) setListViewSize(prefs.listViewSize);
-    // Load divider settings per view
-    if (prefs.fanartHeightByView) {
-      setFanartHeightByView({ ...fanartHeightByView, ...prefs.fanartHeightByView });
-    }
-    if (prefs.descriptionWidthByView) {
-      setDescriptionWidthByView({ ...descriptionWidthByView, ...prefs.descriptionWidthByView });
-    }
-    if (prefs.panelWidthByView) {
-      setPanelWidthByViewState({ ...panelWidthByViewState, ...prefs.panelWidthByView });
-    }
-    // Set initial panelWidth based on current view
-    const savedPanelWidth = (prefs.panelWidthByView && prefs.viewMode ? prefs.panelWidthByView[prefs.viewMode as 'grid' | 'list' | 'logo' | 'carousel' | 'coverflow'] : undefined) ?? prefs.panelWidth;
-    if (savedPanelWidth) setPanelWidth(savedPanelWidth);
-    if (prefs.autoSizeToFit !== undefined) setAutoSizeToFit(prefs.autoSizeToFit);
-    // Restore active game selection if it exists
-    if (prefs.activeGameId) {
-      setActiveGameId(prefs.activeGameId);
-    }
-    if (prefs.isFirstLaunch && baselineDefaultsRef.current) {
-      console.log(`[App] First launch detected. Applying baseline defaults for ${currentResolutionRef.current}.`);
-      applyBaselineDefaults(currentResolutionRef.current);
-      // Save preference change to set isFirstLaunch to false
-      window.electronAPI.savePreferences({ isFirstLaunch: false });
-    }
-
-    if (prefs.confirmGameLaunch !== undefined) setConfirmGameLaunch(prefs.confirmGameLaunch);
-    if (prefs.linkDisplayOrder && prefs.linkDisplayOrder.length > 0) setLinkDisplayOrder(prefs.linkDisplayOrder);
-    if (prefs.visibleLinkTypes && Object.keys(prefs.visibleLinkTypes).length > 0) setVisibleLinkTypes(prefs.visibleLinkTypes);
-
-    // Handle default startup page
-    if (prefs.defaultStartupPage) {
-      if (prefs.defaultStartupPage === 'favorites') {
-        setSelectedCategory('favorites');
-      } else if (prefs.defaultStartupPage === 'recent') {
-        // "Recent" page usually implies sort by last played
-        setSortBy('lastPlayed');
-      } else {
-        // 'library' is default, ensure defaults
-        setSelectedCategory(null);
-      }
-    }
-
-    // Keep current resolution in sync with actual screen height
-    const actualResKey = window.screen.height >= 2160 ? '4K' : window.screen.height >= 1440 ? '1440p' : window.screen.height >= 1080 ? '1080p' : '720p';
-    currentResolutionRef.current = actualResKey;
-    if (prefs.currentResolution !== actualResKey) {
-      window.electronAPI.savePreferences({ currentResolution: actualResKey });
-    }
-
-    if (options?.markInitialLoad) {
-      setIsInitialLoad(false);
-    }
-  };
-
-  // Clamp padding in carousel without overwriting the saved preference
-  const carouselGameTilePadding = (viewMode === 'carousel' || viewMode === 'coverflow') && gameTilePadding > 3 ? 1 : gameTilePadding;
-
-  const refreshPreferences = async () => {
-    const prefs = await window.electronAPI.getPreferences();
-    applyPreferences(prefs);
-  };
-
-  // Load preferences and baseline defaults on mount
-  useEffect(() => {
-    const initialize = async () => {
-      // Load baseline defaults first
-      const baseline = await window.electronAPI.getBaselineDefaults?.();
-      if (baseline) {
-        baselineDefaultsRef.current = baseline;
-      }
-
-      const loadPreferences = async () => {
-        try {
-          const prefs = await window.electronAPI.getPreferences();
-          applyPreferences(prefs, { markInitialLoad: true });
-        } catch (error) {
-          console.error('Error loading preferences:', error);
-          setIsInitialLoad(false);
-        }
-      };
-      await loadPreferences();
-    };
-    initialize();
-  }, []);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    const loadAppInfo = async () => {
-      try {
-        const version = await window.electronAPI.getVersion();
-        if (!cancelled) setCurrentVersion(version);
-      } catch (error) {
-        console.error('Error loading app version:', error);
-      }
-    };
-
-    loadAppInfo();
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const fetchChangelog = useCallback(async (targetVersion?: string) => {
-    if (!window.electronAPI.getChangelog) return;
-    setChangelogLoading(true);
-    setChangelogError(null);
-    setChangelogSource(null);
-
-    try {
-      const result = await window.electronAPI.getChangelog(targetVersion);
-      if (result?.success && result.content) {
-        setChangelogSource(result.content);
-      } else {
-        setChangelogSource(null);
-        setChangelogError(result?.error ?? 'Unable to load changelog.');
-      }
-    } catch (error) {
-      setChangelogSource(null);
-      setChangelogError(error instanceof Error ? error.message : String(error));
-    } finally {
-      setChangelogLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!updateNotification?.version) return;
-    fetchChangelog(updateNotification.version);
-  }, [updateNotification?.version, fetchChangelog]);
-
-  useEffect(() => {
-    if (!import.meta.env.DEV) return;
-    fetchChangelog();
-  }, [fetchChangelog]);
-
-  // Dev-only update popup removed so you can test without the update modal. Use Help > Check for Updates to test the modal.
-
-  const applyBaselineDefaults = (resKey: string) => {
-    if (!baselineDefaultsRef.current || !baselineDefaultsRef.current[resKey]) return;
-    const defaults = baselineDefaultsRef.current[resKey][viewMode];
-    if (!defaults) return;
-
-    // Apply view-specific settings from baseline
-    if (viewMode === 'grid') {
-      if (defaults.gridSize !== undefined) setGridSize(defaults.gridSize);
-      if (defaults.showLogoOverBoxart !== undefined) setShowLogoOverBoxart(defaults.showLogoOverBoxart);
-      if (defaults.gameTilePadding !== undefined) setGameTilePadding(defaults.gameTilePadding);
-    } else if (viewMode === 'logo') {
-      if (defaults.logoSize !== undefined) setLogoSize(defaults.logoSize);
-      if (defaults.logoBackgroundOpacity !== undefined) setLogoBackgroundOpacity(defaults.logoBackgroundOpacity);
-      if (defaults.gameTilePadding !== undefined) setGameTilePadding(defaults.gameTilePadding);
-      if (defaults.rightPanelLogoSize !== undefined) setRightPanelLogoSize(defaults.rightPanelLogoSize);
-    } else if (viewMode === 'list') {
-      if (defaults.listViewOptions !== undefined) setListViewOptions(defaults.listViewOptions);
-      if (defaults.rightPanelLogoSize !== undefined) setRightPanelLogoSize(defaults.rightPanelLogoSize);
-    } else if (viewMode === 'carousel') {
-      if (defaults.showCarouselDetails !== undefined) setShowCarouselDetails(defaults.showCarouselDetails);
-      if (defaults.showCarouselLogos !== undefined) setShowCarouselLogos(defaults.showCarouselLogos);
-      if (defaults.detailsBarSize !== undefined) setDetailsBarSize(defaults.detailsBarSize);
-      if (defaults.selectedBoxArtSize !== undefined) setSelectedBoxArtSize(defaults.selectedBoxArtSize);
-      if (defaults.gameTilePadding !== undefined) setGameTilePadding(defaults.gameTilePadding);
-      if (defaults.carouselLogoSize !== undefined) setCarouselLogoSize(defaults.carouselLogoSize);
-      if (defaults.carouselButtonSize !== undefined) setCarouselButtonSize(defaults.carouselButtonSize);
-      if (defaults.carouselDescriptionSize !== undefined) setCarouselDescriptionSize(defaults.carouselDescriptionSize);
-    } else if (viewMode === 'coverflow') {
-      // Cover Flow uses same-style defaults as carousel where applicable
-      if (defaults && defaults.gameTilePadding !== undefined) setGameTilePadding(defaults.gameTilePadding);
-    }
-
-    // Common baseline settings
-    if (defaults.backgroundBlur !== undefined) setBackgroundBlur(defaults.backgroundBlur);
-    if (defaults.panelWidth !== undefined) setPanelWidth(defaults.panelWidth);
-    if (defaults.rightPanelBoxartPosition !== undefined) setRightPanelBoxartPosition(defaults.rightPanelBoxartPosition);
-    if (defaults.rightPanelBoxartSize !== undefined) setRightPanelBoxartSize(defaults.rightPanelBoxartSize);
-    if (defaults.rightPanelTextSize !== undefined) setRightPanelTextSize(defaults.rightPanelTextSize);
-    if (defaults.rightPanelButtonSize !== undefined) setRightPanelButtonSize(defaults.rightPanelButtonSize);
-    if (defaults.rightPanelButtonLocation !== undefined) setRightPanelButtonLocation(defaults.rightPanelButtonLocation);
-    if (defaults.detailsPanelOpacity !== undefined) setDetailsPanelOpacity(defaults.detailsPanelOpacity);
-  };
-
-  // Detect resolution changes and auto-apply defaults
-  useEffect(() => {
-    const handleResize = () => {
-      const height = window.screen.height;
-      const newResKey = height >= 2160 ? '4K' : height >= 1440 ? '1440p' : height >= 1080 ? '1080p' : '720p';
-
-      if (newResKey !== currentResolutionRef.current) {
-        console.log(`[App] Resolution change detected: ${currentResolutionRef.current} -> ${newResKey}. Auto-applying baseline defaults.`);
-        currentResolutionRef.current = newResKey;
-        applyBaselineDefaults(newResKey);
-        window.electronAPI.savePreferences({ currentResolution: newResKey });
-      }
-    };
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, [viewMode]);
-
-
-  // Save grid size when it changes (but not when auto-size is enabled)
-  useEffect(() => {
-    if (isInitialLoad) return;
-    if (autoSizeToFit) return; // Don't save when auto-size is calculating
-
-    const saveGridSize = async () => {
-      try {
-        await window.electronAPI.savePreferences({ gridSize });
-      } catch (error) {
-        console.error('Error saving grid size:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveGridSize, 500);
-    return () => clearTimeout(timeoutId);
-  }, [gridSize, autoSizeToFit, isInitialLoad]);
-
-  // Save logo size when it changes
-  useEffect(() => {
-    if (isInitialLoad) return;
-    const saveLogoSize = async () => {
-      try {
-        await window.electronAPI.savePreferences({ logoSize });
-      } catch (error) {
-        console.error('Error saving logo size:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveLogoSize, 500);
-    return () => clearTimeout(timeoutId);
-  }, [logoSize, isInitialLoad]);
-
-  // Save pinned categories when they change
-  useEffect(() => {
-    if (isInitialLoad) return;
-    const savePinnedCategories = async () => {
-      try {
-        await window.electronAPI.savePreferences({ pinnedCategories });
-      } catch (error) {
-        console.error('Error saving pinned categories:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(savePinnedCategories, 300);
-    return () => clearTimeout(timeoutId);
-  }, [pinnedCategories, isInitialLoad]);
-
-  // Save hideVRTitles when it changes
-  useEffect(() => {
-    if (isInitialLoad) return;
-    const saveHideVRTitles = async () => {
-      try {
-        await window.electronAPI.savePreferences({ hideVRTitles });
-      } catch (error) {
-        console.error('Error saving hide VR titles preference:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveHideVRTitles, 300);
-    return () => clearTimeout(timeoutId);
-  }, [hideVRTitles, isInitialLoad]);
-
-  // Save hideAppsTitles when it changes
-  useEffect(() => {
-    if (isInitialLoad) return;
-    const saveHideAppsTitles = async () => {
-      try {
-        await window.electronAPI.savePreferences({ hideAppsTitles });
-      } catch (error) {
-        console.error('Error saving hide Apps titles preference:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveHideAppsTitles, 300);
-    return () => clearTimeout(timeoutId);
-  }, [hideAppsTitles, isInitialLoad]);
-
-  // Save appearance preferences when they change (but skip initial load)
-  useEffect(() => {
-    if (isInitialLoad) return; // Skip saving on initial load
-
-    const saveAppearancePrefs = async () => {
-      try {
-        await window.electronAPI.savePreferences({
-          hideGameTitles,
-          gameTilePadding,
-          backgroundBlur,
-          backgroundBrightnessByView,
-          viewMode,
-          backgroundMode,
-          backgroundColor,
-          listViewOptions,
-          listViewSize
-        });
-      } catch (error) {
-        console.error('Error saving appearance preferences:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveAppearancePrefs, 500);
-    return () => clearTimeout(timeoutId);
-  }, [hideGameTitles, gameTilePadding, backgroundBlur, backgroundBrightnessByView, viewMode, isInitialLoad]);
-
-  // Save activeGameId when it changes
-  useEffect(() => {
-    if (isInitialLoad) return;
-    const saveActiveGameId = async () => {
-      try {
-        await window.electronAPI.savePreferences({ activeGameId });
-      } catch (error) {
-        console.error('Error saving active game ID:', error);
-      }
-    };
-    // Debounce saves
-    const timeoutId = setTimeout(saveActiveGameId, 300);
-    return () => clearTimeout(timeoutId);
-  }, [activeGameId, isInitialLoad]);
-
-  // Restore active game selection after games are loaded
-  useEffect(() => {
-    if (!loading && games.length > 0 && activeGameId) {
-      // Verify the saved game still exists in the library
-      const gameExists = games.some(g => g.id === activeGameId);
-      if (!gameExists) {
-        // Game no longer exists, clear the selection
-        setActiveGameId(null);
-      }
-    }
-  }, [loading, games, activeGameId]);
-
-  // Toggle pin category
-  const handleTogglePinCategory = (category: string) => {
-    setPinnedCategories(prev => {
-      if (prev.includes(category)) {
-        return prev.filter(c => c !== category);
-      } else {
-        return [...prev, category];
-      }
-    });
-  };
-
-  // Filter and sort state
-  const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
-  const [selectedLauncher, setSelectedLauncher] = useState<string | null>(null);
-  const [sortBy, setSortBy] = useState<'title' | 'releaseDate' | 'playtime' | 'lastPlayed'>('title');
-
-
-  // Listen to menu events
-  useEffect(() => {
-    const cleanup1 = window.electronAPI.onMenuEvent('menu:addGame', async () => {
-      const apisConfigured = await areAPIsConfigured();
-      if (!apisConfigured) {
-        showToast('Both IGDB (Client ID + Secret) and SteamGridDB (API Key) are required before adding games. Please configure them in Settings > APIs.', 'error');
-        setIsOnyxSettingsOpen(true);
-        setOnyxSettingsInitialTab('apis');
-        return;
-      }
-      setIsModalOpen(true);
-    });
-    const cleanup2 = window.electronAPI.onMenuEvent('menu:scanFolder', () => {
-      handleScanFolder();
-    });
-    const cleanup3 = window.electronAPI.onMenuEvent('menu:updateSteamLibrary', () => {
-      handleUpdateSteamLibrary();
-    });
-    const cleanup4 = window.electronAPI.onMenuEvent('menu:configureSteam', () => {
-      setIsSteamConfigOpen(true);
-    });
-    const cleanup5 = window.electronAPI.onMenuEvent('menu:checkForUpdates', () => {
-      window.electronAPI.checkForUpdates?.();
-    });
-
-    // Listen for new Steam games found notification
-    const newGamesHandler = (_event: any, data: { count: number; games: Array<any> }) => {
-      // Always use FoundGamesModal for consistency
-      if (data.games && data.games.length > 0) {
-        setFoundGames(data.games);
-      }
-    };
-
-    // Listen for new games found from background scan (all sources)
-    const backgroundNewGamesHandler = (_event: any, data: { count: number; games: Array<any>; bySource?: Record<string, Array<any>> }) => {
-      console.log('[App] Background scan found new games:', data);
-      if (data.games && data.games.length > 0) {
-        setFoundGames(data.games);
-      }
-    };
-
-    // Listen for startup scan progress
-    const startupProgressHandler = (_event: any, data: { message: string }) => {
-      // Handle cases where data might be undefined or malformed
-      if (!data || typeof data !== 'object') {
-        console.warn('[App] ⚠️ Received malformed startup:progress data:', data);
-        return;
-      }
-      setStartupProgress(data);
-      // Auto-hide progress after completion, but keep open if games were found for user interaction
-      if (data.message && (data.message.includes('Scan complete') || data.message.includes('Error'))) {
-        setTimeout(() => {
-          setFoundGames(currentFoundGames => {
-            // If games were found, keep the overlay open (don't close startupProgress)
-            if (currentFoundGames && currentFoundGames.length > 0) {
-              return currentFoundGames;
-            }
-            // No games found, close the overlay after 500ms
-            setStartupProgress(null);
-            return currentFoundGames;
-          });
-        }, 500);
-      }
-    };
-
-    // Listen for missing games detected during scans
-    const missingGamesHandler = (_event: any, data: { games: Array<any> }) => {
-      console.log('[App] Missing games detected:', data);
-      if (data.games && data.games.length > 0) {
-        setMissingGames(data.games);
-      }
-    };
-
-    const removeSteamNewGames = window.electronAPI?.on && window.electronAPI.on('steam:newGamesFound', newGamesHandler);
-    const removeBackgroundNewGames = window.electronAPI?.on && window.electronAPI.on('background:newGamesFound', backgroundNewGamesHandler);
-    const removeStartupProgress = window.electronAPI?.on && window.electronAPI.on('startup:progress', startupProgressHandler);
-    const removeMissingGames = window.electronAPI?.on && window.electronAPI.on('scan:missing-games', missingGamesHandler);
-
-    // Listen for update status: show persistent notification when update is available
-    const updateStatusHandler = (_event: any, payload: { status: string; version?: string; error?: string }) => {
-      if (payload.status === 'available' && payload.version) {
-        setIsUpdateModalTest(false);
-        setUpdateNotification({
-          version: payload.version,
-          status: 'available',
-        });
-        // Signal main process that update was found (to pause startup scan)
-        window.electronAPI.onUpdateFound?.();
-      } else if (payload.status === 'downloading') {
-        setIsUpdateModalTest(false);
-        setUpdateNotification(prev => prev ? { ...prev, status: 'downloading' } : null);
-      } else if (payload.status === 'downloaded') {
-        setIsUpdateModalTest(false);
-        setUpdateNotification(prev => prev ? { ...prev, status: 'downloaded' } : null);
-      } else if (payload.status === 'error' && payload.error) {
-        setIsUpdateModalTest(false);
-        setUpdateNotification(prev => prev ? { ...prev, status: 'error', error: payload.error } : null);
-      }
-    };
-    const removeUpdateStatus = window.electronAPI?.on && window.electronAPI.on('app:update-status', updateStatusHandler);
-
-    return () => {
-      cleanup1();
-      cleanup2();
-      cleanup3();
-      cleanup4();
-      cleanup5();
-      if (typeof removeSteamNewGames === 'function') removeSteamNewGames();
-      if (typeof removeBackgroundNewGames === 'function') removeBackgroundNewGames();
-      if (typeof removeStartupProgress === 'function') removeStartupProgress();
-      if (typeof removeMissingGames === 'function') removeMissingGames();
-      if (typeof removeUpdateStatus === 'function') removeUpdateStatus();
-    };
-  }, []);
-
-  // Get all unique categories and their counts from games
-  const { allCategories, categoryCounts } = useMemo(() => {
-    const counts: Record<string, number> = {};
-    let favorites = 0;
-    let hidden = 0;
-    games.forEach(game => {
-      if (game.favorite) favorites++;
-      if (game.hidden) hidden++;
-      game.categories?.forEach(cat => {
-        counts[cat] = (counts[cat] || 0) + 1;
-      });
-    });
-
-    return {
-      allCategories: Object.keys(counts).sort(),
-      categoryCounts: {
-        ...counts,
-        favorites,
-        hidden,
-        all: games.length
-      }
-    };
-  }, [games]);
-
-  // Automatically pin VR and Apps categories when they exist
-  useEffect(() => {
-    const categoriesToAutoPin = ['VR', 'Apps'];
-    setPinnedCategories(prev => {
-      const updated = [...prev];
-      let changed = false;
-
-      categoriesToAutoPin.forEach(category => {
-        if (allCategories.includes(category) && !prev.includes(category)) {
-          updated.push(category);
-          changed = true;
-        }
-      });
-
-      return changed ? updated : prev;
-    });
-  }, [allCategories]);
-
-  // Check if there are any favorite games
-  const hasFavoriteGames = useMemo(() => {
-    return games.some(g => g.favorite === true);
-  }, [games]);
-
-  // Check if VR category exists
-  const hasVRCategory = useMemo(() => {
-    return allCategories.includes('VR');
-  }, [allCategories]);
-
-  // Check if Apps category exists
-  const hasAppsCategory = useMemo(() => {
-    return allCategories.includes('Apps');
-  }, [allCategories]);
-
-  // Check if there are any hidden games
-  const hasHiddenGames = useMemo(() => {
-    return games.some(g => g.hidden === true);
-  }, [games]);
-
-  // Get launcher from game (check ID format, then source, then platform, then installation directory)
-  const getGameLauncher = useCallback((game: Game): string => {
-    // Check ID format first (most reliable)
-    if (game.id.startsWith('steam-')) {
-      return 'steam';
-    }
-    if (game.id.startsWith('epic-')) {
-      return 'epic';
-    }
-    if (game.id.startsWith('gog-')) {
-      return 'gog';
-    }
-    if (game.id.startsWith('xbox-')) {
-      return 'xbox';
-    }
-    if (game.id.startsWith('ubisoft-')) {
-      return 'ubisoft';
-    }
-    if (game.id.startsWith('rockstar-')) {
-      return 'rockstar';
-    }
-    if (game.id.startsWith('ea-') || game.id.startsWith('origin-')) {
-      return 'ea';
-    }
-    if (game.id.startsWith('battle-') || game.id.startsWith('battlenet-')) {
-      return 'battle';
-    }
-
-    // Check source field
-    if (game.source) {
-      const source = game.source.toLowerCase();
-      const validSources = ['steam', 'epic', 'gog', 'xbox', 'ea', 'origin', 'ubisoft', 'battle', 'battlenet', 'humble', 'itch', 'rockstar'];
-      if (validSources.includes(source)) {
-        // Normalize some source names
-        if (source === 'origin') return 'ea';
-        if (source === 'battlenet') return 'battle';
-        return source;
-      }
-    }
-
-    // Check platform field (fallback)
-    const platform = game.platform?.toLowerCase();
-    if (platform === 'steam') {
-      return 'steam';
-    }
-    if (platform === 'epic' || platform === 'epic games') {
-      return 'epic';
-    }
-    if (platform === 'gog' || platform === 'gog galaxy') {
-      return 'gog';
-    }
-    if (platform === 'xbox' || platform === 'xbox game pass') {
-      return 'xbox';
-    }
-    if (platform === 'ea' || platform === 'ea app' || platform === 'origin') {
-      return 'ea';
-    }
-    if (platform === 'ubisoft' || platform === 'ubisoft connect') {
-      return 'ubisoft';
-    }
-    if (platform === 'battle.net' || platform === 'battlenet' || platform === 'battle') {
-      return 'battle';
-    }
-    if (platform === 'rockstar' || platform === 'rockstar games') {
-      return 'rockstar';
-    }
-
-    // Check installation directory as last resort
-    if (game.installationDirectory) {
-      const installPath = game.installationDirectory.toLowerCase();
-      if (installPath.includes('steam')) return 'steam';
-      if (installPath.includes('epic games') || installPath.includes('epicgames')) return 'epic';
-      if (installPath.includes('gog galaxy') || installPath.includes('gog\\games')) return 'gog';
-      if (installPath.includes('xboxgames') || installPath.includes('windowsapps')) return 'xbox';
-      if (installPath.includes('electronic arts') || installPath.includes('ea games') || installPath.includes('origin')) return 'ea';
-      if (installPath.includes('ubisoft')) return 'ubisoft';
-      if (installPath.includes('battle.net') || installPath.includes('battlenet')) return 'battle';
-      if (installPath.includes('rockstar games')) return 'rockstar';
-      if (installPath.includes('humble')) return 'humble';
-      if (installPath.includes('itch')) return 'itch';
-    }
-
-    return 'other';
-  }, []);
-
-  // Get all unique launchers from games
-  const allLaunchers = useMemo(() => {
-    const launchers = new Set<string>();
-    games.forEach(game => {
-      const launcher = getGameLauncher(game);
-      if (launcher) {
-        launchers.add(launcher);
-      }
-    });
-    return Array.from(launchers).sort((a, b) => {
-      // Sort with 'other' at the end
-      if (a === 'other') return 1;
-      if (b === 'other') return -1;
-      return a.localeCompare(b);
-    });
-  }, [games, getGameLauncher]);
-
-  // Filter games based on search, section, and category
-  const filteredGames = useMemo(() => {
-    let filtered = games;
-
-    // Filter by section
-    if (activeSection === 'favorites') {
-      filtered = filtered.filter(g => g.favorite);
-    } else if (activeSection === 'recent') {
-      filtered = filtered.filter(g => g.lastPlayed);
-    }
-
-    // Filter by category or favorites
-    if (selectedCategory === 'favorites') {
-      filtered = filtered.filter(g => g.favorite === true);
-    } else if (selectedCategory === 'hidden') {
-      // Show only hidden games when "Hidden" category is selected
-      filtered = filtered.filter(g => g.hidden === true);
-    } else if (selectedCategory) {
-      filtered = filtered.filter(g =>
-        g.categories?.includes(selectedCategory)
-      );
-    }
-
-    // Filter out hidden games by default (unless "Hidden" category is selected)
-    if (selectedCategory !== 'hidden') {
-      filtered = filtered.filter(g => g.hidden !== true);
-    }
-
-    // Filter by launcher
-    if (selectedLauncher) {
-      filtered = filtered.filter(g => {
-        const gameLauncher = getGameLauncher(g);
-        return gameLauncher === selectedLauncher;
-      });
-    }
-
-    // Filter out VR titles if hideVRTitles is enabled, but not if VR category is selected
-    if (hideVRTitles && selectedCategory !== 'VR' && selectedCategory !== 'Apps') {
-      filtered = filtered.filter(g =>
-        !g.categories?.includes('VR')
-      );
-    }
-
-    // Filter out Apps titles if hideAppsTitles is enabled, but not if Apps category is selected
-    if (hideAppsTitles && selectedCategory !== 'Apps' && selectedCategory !== 'VR') {
-      filtered = filtered.filter(g =>
-        !g.categories?.includes('Apps')
-      );
-    }
-
-    // Filter by search query
-    if (searchQuery.trim()) {
-      const query = searchQuery.toLowerCase();
-      filtered = filtered.filter(g =>
-        g.title.toLowerCase().includes(query) ||
-        g.genres?.some(genre => genre.toLowerCase().includes(query)) ||
-        g.developers?.some(dev => dev.toLowerCase().includes(query))
-      );
-    }
-
-    // Sort games - pinned games always appear first
-    filtered = [...filtered].sort((a, b) => {
-      // First, sort by pinned status (pinned games first)
-      const aPinned = a.pinned === true ? 1 : 0;
-      const bPinned = b.pinned === true ? 1 : 0;
-      if (aPinned !== bPinned) {
-        return bPinned - aPinned; // Pinned games first
-      }
-
-      // Then sort by the selected criteria
-      switch (sortBy) {
-        case 'title':
-          return (a.sortingName || a.title).localeCompare(b.sortingName || b.title);
-        case 'releaseDate':
-          const dateA = a.releaseDate ? new Date(a.releaseDate).getTime() : 0;
-          const dateB = b.releaseDate ? new Date(b.releaseDate).getTime() : 0;
-          return dateB - dateA; // Newest first
-        case 'playtime':
-          const playtimeA = a.playtime || 0;
-          const playtimeB = b.playtime || 0;
-          return playtimeB - playtimeA; // Most played first
-        case 'lastPlayed':
-          const lastA = a.lastPlayed ? new Date(a.lastPlayed).getTime() : 0;
-          const lastB = b.lastPlayed ? new Date(b.lastPlayed).getTime() : 0;
-          return lastB - lastA; // Most recent first
-        default:
-          return 0;
-      }
-    });
-
-    return filtered;
-  }, [games, searchQuery, activeSection, selectedCategory, selectedLauncher, sortBy, hideVRTitles, hideAppsTitles]);
-
-  const activeGame = activeGameId ? games.find(g => g.id === activeGameId) || null : null;
-
-  // When no game is selected, select the first (top-left) game in the current view
-  useEffect(() => {
-    if (!loading && filteredGames.length > 0 && !activeGameId) {
-      setActiveGameId(filteredGames[0].id);
-    }
-  }, [loading, filteredGames, activeGameId]);
-
-  const handlePlay = async (game: Game) => {
-    if (confirmGameLaunch) {
-      setLaunchConfirmation({ game });
-      return;
-    }
-
-    launchGame(game);
-  };
-
-  const launchGame = async (game: Game) => {
-    setLaunchingGameId(game.id);
-    try {
-      const result = await window.electronAPI.launchGame(game.id);
-      if (!result.success) {
-        console.error('Failed to launch game:', result.error);
-        alert(`Failed to launch game: ${result.error || 'Unknown error'}`);
-        setLaunchingGameId(null);
-        return;
-      }
-
-      // Check if we should minimize the window on game launch
-      const prefs = await window.electronAPI.getPreferences();
-      if (prefs.minimizeOnGameLaunch) {
-        await window.electronAPI.minimizeWindow();
-      }
-
-      // Notify main process that a game has started
-      await window.electronAPI.scanning?.gameStarted?.(game.id);
-
-      // Game launched successfully
-      // Wait a moment for the process to start, then mark as running
-      setTimeout(() => {
-        setLaunchingGameId(null);
-        setRunningGames(prev => new Set(prev).add(game.id));
-
-        // For non-Steam games with PIDs, monitor the process
-        if (result.pid) {
-          monitorGameProcess(game.id, result.pid);
-        } else {
-          // For Steam games or games without PID, poll for process
-          pollForGameProcess(game.id);
-        }
-      }, 1000); // Show "Launching..." for 1 second
-
-    } catch (error) {
-      console.error('Error launching game:', error);
-      alert(`Error launching game: ${error instanceof Error ? error.message : 'Unknown error'}`);
-      setLaunchingGameId(null);
-    }
-  };
-
-  // Monitor game process and update state when it closes
-  const monitorGameProcess = (gameId: string, pid: number) => {
-    const checkInterval = setInterval(async () => {
-      try {
-        // Check if process is still running by trying to kill with signal 0
-        // This doesn't actually kill the process, just checks if it exists
-        const isRunning = await checkProcessRunning(pid);
-        if (!isRunning) {
-          clearInterval(checkInterval);
-          setRunningGames(prev => {
-            const newSet = new Set(prev);
-            newSet.delete(gameId);
-            return newSet;
-          });
-
-          // Notify main process that a game has stopped
-          await window.electronAPI.scanning?.gameStopped?.(gameId);
-
-          // Check if we should restore the window on game exit
-          const prefs = await window.electronAPI.getPreferences();
-          if (prefs.restoreAfterLaunch) {
-            const { isMinimized } = await window.electronAPI.fullscreen.isMinimized();
-            if (isMinimized) {
-              await window.electronAPI.restoreWindow();
-            }
-          }
-        }
-      } catch (error) {
-        console.error('Error checking process:', error);
-        clearInterval(checkInterval);
-      }
-    }, 2000); // Check every 2 seconds
-  };
-
-  // Poll for game process for Steam games
-  const pollForGameProcess = (gameId: string) => {
-    let pollCount = 0;
-    const maxPolls = 30; // Poll for up to 60 seconds (30 * 2s)
-
-    const checkInterval = setInterval(async () => {
-      pollCount++;
-
-      // After max polls, assume game closed
-      if (pollCount > maxPolls) {
-        clearInterval(checkInterval);
-        setRunningGames(prev => {
-          const newSet = new Set(prev);
-          newSet.delete(gameId);
-          return newSet;
-        });
-
-        // Notify main process that a game has stopped
-        await window.electronAPI.scanning?.gameStopped?.(gameId);
-
-        // Check if we should restore the window on game exit
-        const prefs = await window.electronAPI.getPreferences();
-        if (prefs.restoreAfterLaunch) {
-          const { isMinimized } = await window.electronAPI.fullscreen.isMinimized();
-          if (isMinimized) {
-            await window.electronAPI.restoreWindow();
-          }
-        }
-        return;
-      }
-
-      // For now, just keep it running for a reasonable time
-      // TODO: Implement actual process checking for Steam games
-    }, 2000);
-  };
-
-  // Helper to check if process is running
-  const checkProcessRunning = async (pid: number): Promise<boolean> => {
-    try {
-      // On Windows, we can use tasklist to check if process exists
-      const result = await window.electronAPI.checkProcessExists(pid);
-      return result ?? false;
-    } catch {
-      return false;
-    }
-  };
-
-  const handleReorder = async (reorderedGames: Game[]) => {
-    await reorderGames(reorderedGames);
-  };
-
-  const calculateAutoSize = useCallback(() => {
-    if (!gridContainerRef.current || viewMode !== 'grid' || filteredGames.length === 0) {
-      return;
-    }
-
-    const container = gridContainerRef.current;
-    // Use the actual container width (left panel) to ensure we fill it properly
-    const containerWidth = container.clientWidth;
-    // Use the visible viewport height, not the scrollable container height
-    const containerHeight = container.clientHeight; // This is the visible height
-
-    // Account for padding (p-4 = 16px on each side = 32px total)
-    const horizontalPadding = 32;
-    const verticalPadding = 32; // Top and bottom padding
-    const availableWidth = containerWidth - horizontalPadding;
-    const availableHeight = containerHeight - verticalPadding;
-
-    if (availableWidth <= 0 || availableHeight <= 0) {
-      return;
-    }
-
-    const totalGames = filteredGames.length;
-    const gap = gameTilePadding;
-
-    // GameCard uses aspect-[2/3], so height = width * 1.5
-    // We need to find the grid size that maximizes tile size while ensuring
-    // the rightmost boxart gets as close as possible to the divider
-
-    let bestSize = 0;
-
-    let bestRemainingWidth = Infinity;
-
-    // Try different column counts to find the one that fills the width best
-    for (let columns = 1; columns <= 20; columns++) {
-      // Calculate tile width based on available width
-      const totalGapWidth = gap * (columns - 1);
-      const tileWidth = (availableWidth - totalGapWidth) / columns;
-
-      if (tileWidth < 50) continue; // Too small, skip
-
-      // Calculate tile height (2:3 aspect ratio)
-      const tileHeight = tileWidth * 1.5;
-
-      // Calculate how many rows we need to fit all games
-      const rowsNeeded = Math.ceil(totalGames / columns);
-
-      // Calculate total height needed for all rows
-      const totalHeightNeeded = (tileHeight * rowsNeeded) + (gap * (rowsNeeded - 1));
-
-      // Check if this configuration fits ALL games in the visible height
-      if (totalHeightNeeded <= availableHeight) {
-        // Calculate how much space this configuration uses
-        const usedWidth = (tileWidth * columns) + (gap * (columns - 1));
-        const remainingWidth = availableWidth - usedWidth;
-
-        // Prioritize configurations that minimize remaining width (fill more space)
-        // Among those, prefer larger tile sizes
-        if (bestSize === 0 ||
-          remainingWidth < bestRemainingWidth ||
-          (Math.abs(remainingWidth - bestRemainingWidth) < 5 && tileWidth > bestSize)) {
-          bestSize = tileWidth;
-
-          bestRemainingWidth = remainingWidth;
-        }
-      }
-    }
-
-    // If we found a solution, use it
-    if (bestSize > 0) {
-      setGridSize(Math.round(bestSize));
-    } else {
-      // No solution found - try to fit as many as possible
-      // Start with a reasonable tile size and work backwards
-      for (let testSize = 200; testSize >= 50; testSize -= 10) {
-        const tileHeight = testSize * 1.5;
-
-        for (let columns = 1; columns <= 20; columns++) {
-          const totalGapWidth = gap * (columns - 1);
-          const tileWidth = (availableWidth - totalGapWidth) / columns;
-
-          if (Math.abs(tileWidth - testSize) < 10) { // Close match
-            const rowsNeeded = Math.ceil(totalGames / columns);
-            const totalHeightNeeded = (tileHeight * rowsNeeded) + (gap * (rowsNeeded - 1));
-
-            if (totalHeightNeeded <= availableHeight) {
-              setGridSize(Math.round(tileWidth));
-              return;
-            }
-          }
-        }
-      }
-
-      // Last resort: use a small size that should fit
-      const minColumns = Math.ceil(Math.sqrt(totalGames));
-      const totalGapWidth = gap * (minColumns - 1);
-      const fallbackSize = Math.round((availableWidth - totalGapWidth) / minColumns);
-      setGridSize(Math.max(50, Math.min(500, fallbackSize)));
-    }
-  }, [viewMode, filteredGames.length, gameTilePadding, hideGameTitles]);
-
-  // Auto-recalculate when auto-size is enabled and dependencies change
-  useEffect(() => {
-    if (!autoSizeToFit || viewMode !== 'grid' || filteredGames.length === 0) {
-      return;
-    }
-
-    // Recalculate after a short delay to ensure DOM is updated
-    const timeoutId = setTimeout(() => {
-      calculateAutoSize();
-    }, 100);
-
-    return () => clearTimeout(timeoutId);
-  }, [autoSizeToFit, filteredGames.length, gameTilePadding, hideGameTitles, viewMode, calculateAutoSize, panelWidth]);
-
-  // Watch for container size changes using ResizeObserver (handles window resize and panel resize)
-  useEffect(() => {
-    if (!autoSizeToFit || viewMode !== 'grid' || !gridContainerRef.current) {
-      return;
-    }
-
-    const container = gridContainerRef.current;
-    let resizeTimeout: NodeJS.Timeout;
-
-    const resizeObserver = new ResizeObserver(() => {
-      // Debounce resize calculations
-      clearTimeout(resizeTimeout);
-      resizeTimeout = setTimeout(() => {
-        if (filteredGames.length > 0) {
-          calculateAutoSize();
-        }
-      }, 150);
-    });
-
-    resizeObserver.observe(container);
-
-    // Also listen to window resize as a fallback
-    const handleWindowResize = () => {
-      clearTimeout(resizeTimeout);
-      resizeTimeout = setTimeout(() => {
-        if (filteredGames.length > 0) {
-          calculateAutoSize();
-        }
-      }, 150);
-    };
-
-    window.addEventListener('resize', handleWindowResize);
-
-    return () => {
-      resizeObserver.disconnect();
-      window.removeEventListener('resize', handleWindowResize);
-      clearTimeout(resizeTimeout);
-    };
-  }, [autoSizeToFit, viewMode, filteredGames.length, calculateAutoSize]);
-
-
-  const handleGameClick = (game: Game) => {
-    setActiveGameId(game.id);
-  };
-
-  const handleEditGame = (game: Game) => {
-    setGameManagerInitialGameId(game.id);
-    setGameManagerInitialTab('metadata');
-    setIsGameManagerOpen(true);
-  };
-
-  const handleEditCategories = (game: Game) => {
-    setEditingCategoriesGame(game);
-    setIsCategoriesEditorOpen(true);
-  };
-
-  const handleEditImages = (game: Game) => {
-    setGameManagerInitialGameId(game.id);
-    setGameManagerInitialTab('images');
-    setIsGameManagerOpen(true);
-  };
-
-  const handleFixMatch = (game: Game) => {
-    setGameManagerInitialGameId(game.id);
-    setGameManagerInitialTab('metadata');
-    setIsGameManagerOpen(true);
-  };
-
-  const handleSelectMetadataMatch = async (result: { id: string; source: string }) => {
-    if (!fixingGame) return;
-
-    try {
-      const response = await window.electronAPI.fetchAndUpdateByProviderId(
-        fixingGame.id,
-        result.id,
-        result.source
-      );
-
-      if (response.success) {
-        showToast('Metadata updated successfully!', 'success');
-        // Reload library to show updated metadata
-        await loadLibrary();
-      } else {
-        showToast(response.error || 'Failed to update metadata', 'error');
-      }
-    } catch (error) {
-      console.error('Error updating metadata:', error);
-      showToast('An error occurred while updating metadata', 'error');
-    }
-  };
-
-  const handleSaveGame = async (game: Game, oldGame?: Game) => {
-    try {
-      console.log('Saving game from App:', game.title, 'favorite:', game.favorite);
-      // Get old game if not provided
-      if (!oldGame) {
-        oldGame = games.find(g => g.id === game.id);
-      }
-      const success = await window.electronAPI.saveGame(game, oldGame);
-      if (success) {
-        await loadLibrary();
-        // Verify the game was saved correctly
-        const updatedGames = await window.electronAPI.getLibrary();
-        const savedGame = updatedGames.find(g => g.id === game.id);
-        console.log('Game after save - favorite:', savedGame?.favorite, 'Full game:', savedGame);
-        showToast(`Game "${game.title}" updated successfully`, 'success');
-      } else {
-        showToast('Failed to save game', 'error');
-      }
-    } catch (err) {
-      console.error('Error in handleSaveGame:', err);
-      showToast('Failed to save game', 'error');
-    }
-  };
-
-  const handleAddFolder = async (path: string, categories: string[]) => {
-    try {
-      // Create config with default name (folder basename)
-      const folderName = path.split(/[/\\]/).pop() || 'Manual Folder';
-      // Generate a simple ID from the path
-      const pathHash = btoa(path).replace(/[^a-zA-Z0-9]/g, '').substring(0, 16);
-      const folderId = `manual-${pathHash}`;
-
-      const newConfig = {
-        id: folderId,
-        name: folderName,
-        path: path,
-        enabled: true,
-        autoCategory: categories
-      };
-
-      if (window.electronAPI.saveManualFolderConfig) {
-        await window.electronAPI.saveManualFolderConfig(newConfig);
-        showToast(`Successfully added folder: ${folderName}`, 'success');
-        // Refresh library to pick up any games in the new folder
-        loadLibrary();
-      }
-    } catch (err) {
-      console.error('Error adding manual folder:', err);
-      showToast('Failed to add folder', 'error');
-    }
-  };
-
-
-  const handleAddGame = async (game: Game) => {
-    // Check if APIs are configured
-    const apisConfigured = await areAPIsConfigured();
-    if (!apisConfigured) {
-      showToast('API credentials must be configured before adding games. Please configure them in Settings.', 'error');
-      setIsModalOpen(false);
-      setIsOnyxSettingsOpen(true);
-      setOnyxSettingsInitialTab('apis');
-      return;
-    }
-    await addCustomGame(game);
-  };
-
-  // Toast notification helper
-  const showToast = (message: string, type: 'success' | 'error' = 'success') => {
-    setToast({ message, type });
-    setTimeout(() => setToast(null), 3000);
-  };
-
-  // Missing games handlers
-  const handleRemoveMissingGames = async (gameIds: string[]) => {
-    try {
-      const result = await window.electronAPI.removeMissingGames(gameIds);
-      if (result.success) {
-        showToast(`Successfully removed ${result.removedCount} missing game(s)`, 'success');
-        // Reload library to reflect changes
-        await loadLibrary();
-      } else {
-        showToast('Failed to remove missing games', 'error');
-      }
-    } catch (error) {
-      console.error('Error removing missing games:', error);
-      showToast('Failed to remove missing games', 'error');
-    } finally {
-      setMissingGames(null);
-    }
-  };
-
-  const handleCancelMissingGames = () => {
-    setMissingGames(null);
-  };
-
-  // Hide confirmation dialog state
-  const [hideConfirmation, setHideConfirmation] = useState<{ game: Game } | null>(null);
-
-  // Update Library handler - opens game importer
-  const handleUpdateSteamLibrary = async () => {
-    // Check if APIs are configured
-    const apisConfigured = await areAPIsConfigured();
-    if (!apisConfigured) {
-      showToast('API credentials must be configured before adding games. Please configure them in Settings.', 'error');
-      setIsOnyxSettingsOpen(true);
-      setOnyxSettingsInitialTab('apis');
-      return;
-    }
-    setIsImportWorkbenchOpen(true);
-  };
-
-  // Handle Steam games import
-
-  // Handle Steam configuration scan
-  const handleSteamConfigScan = async (steamPath?: string) => {
-    setIsScanningSteam(true);
-
-    try {
-      const beforeCount = games.length;
-      const result = await window.electronAPI.scanGamesWithPath(steamPath);
-
-      if (result.success) {
-        const updatedGames = await window.electronAPI.getLibrary();
-        const afterCount = updatedGames.length;
-        const newGamesCount = afterCount - beforeCount;
-
-        // Reload library to update UI
-        await loadLibrary();
-
-        if (newGamesCount > 0) {
-          showToast(`Library updated: ${newGamesCount} new ${newGamesCount === 1 ? 'game' : 'games'} found`, 'success');
-        } else {
-          showToast('Steam library is up to date', 'success');
-        }
-      } else {
-        throw new Error(result.error || 'Failed to scan Steam library');
-      }
-    } catch (err) {
-      console.error('Error scanning Steam library:', err);
-      throw err; // Re-throw to let modal handle it
-    } finally {
-      setIsScanningSteam(false);
-    }
-  };
-
-  // Scan Folder handler - now uses ImportWorkbench
-  const handleScanFolder = async () => {
-    // Check if APIs are configured
-    const apisConfigured = await areAPIsConfigured();
-    if (!apisConfigured) {
-      showToast('API credentials must be configured before adding games. Please configure them in Settings.', 'error');
-      setIsOnyxSettingsOpen(true);
-      setOnyxSettingsInitialTab('apis');
-      return;
-    }
-
-    try {
-      const folderPath = await window.electronAPI.showFolderDialog();
-      if (!folderPath) {
-        return; // User cancelled
-      }
-
-      // Open ImportWorkbench with the selected folder path
-      setImportWorkbenchFolderPath(folderPath);
-      setIsImportWorkbenchOpen(true);
-    } catch (err) {
-      console.error('Error selecting folder:', err);
-      showToast('Failed to select folder', 'error');
-    }
-  };
-
-
-  // Handle save from metadata editor
-  const handleSaveGameWithMetadata = async (title: string, exePath: string, metadata: GameMetadata) => {
-    try {
-      // Create game with all metadata
-      const gameId = `custom-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-      const newGame: Game = {
-        id: gameId,
-        title: metadata.title || title,
-        platform: metadata.platform || 'other',
-        exePath,
-        boxArtUrl: metadata.boxArtUrl || '',
-        bannerUrl: metadata.bannerUrl || '',
-        description: metadata.description,
-        releaseDate: metadata.releaseDate,
-        genres: metadata.genres,
-        ageRating: metadata.ageRating,
-        categories: metadata.categories,
-        useAlternativeBackground: true,
-      };
-
-      // Save the game
-      const success = await window.electronAPI.saveGame(newGame);
-      if (success) {
-        await loadLibrary();
-        showToast(`Game "${metadata.title || title}" added successfully`, 'success');
-        setIsMetadataEditorOpen(false);
-        setSelectedExecutable(null);
-      } else {
-        showToast('Failed to save game', 'error');
-      }
-    } catch (err) {
-      console.error('Error saving game with metadata:', err);
-      showToast('Failed to save game', 'error');
-    }
-  };
-
-  const handleToggleFavorite = async (game: Game) => {
-    const newFavoriteValue = game.favorite !== true; // Explicitly set to true or false
-    const updatedGame = { ...game, favorite: newFavoriteValue };
-    console.log('Toggling favorite for game:', game.title, 'Current favorite:', game.favorite, 'New favorite value:', newFavoriteValue);
-    await handleSaveGame(updatedGame);
-  };
-
-  const handleTogglePin = async (game: Game) => {
-    const newPinnedValue = game.pinned !== true; // Explicitly set to true or false
-    const updatedGame = { ...game, pinned: newPinnedValue };
-    await handleSaveGame(updatedGame);
-  };
-
-  const handleHideGame = (game: Game) => {
-    setHideConfirmation({ game });
-  };
-
-  const handleConfirmHide = async () => {
-    if (hideConfirmation) {
-      const { game } = hideConfirmation;
-      const updatedGame = { ...game, hidden: true };
-      await handleSaveGame(updatedGame);
-      showToast(`"${game.title}" has been hidden`, 'success');
-      setHideConfirmation(null);
-    }
-  };
-
-  const handleCancelHide = () => {
-    setHideConfirmation(null);
-  };
-
-  const handleUnhideGame = async (game: Game) => {
-    const updatedGame = { ...game, hidden: false };
-    await handleSaveGame(updatedGame);
-    showToast(`"${game.title}" has been unhidden`, 'success');
-  };
-
-  // Handle exit with confirmation
-  const handleExit = async () => {
-    try {
-      const exitInfo = await window.electronAPI.requestExit();
-
-      if (exitInfo.shouldMinimizeToTray && exitInfo.canMinimizeToTray) {
-        // Show confirmation dialog asking if user wants to minimize to tray instead
-        const shouldMinimize = window.confirm(
-          'Do you want to minimize Onyx to the system tray instead of exiting?\n\n' +
-          'Click OK to minimize to tray, or Cancel to exit.'
-        );
-
-        if (shouldMinimize) {
-          await window.electronAPI.minimizeToTray();
-        } else {
-          await window.electronAPI.exit();
-        }
-      } else {
-        // Show confirmation dialog
-        const shouldExit = window.confirm('Are you sure you want to exit Onyx?');
-        if (shouldExit) {
-          await window.electronAPI.exit();
-        }
-      }
-    } catch (error) {
-      console.error('Error handling exit:', error);
-      // Fallback to simple confirmation
-      const shouldExit = window.confirm('Are you sure you want to exit Onyx?');
-      if (shouldExit) {
-        await window.electronAPI.exit();
-      }
-    }
-  };
-
-  // Get background image from active game - use alternative banner if enabled
-  const backgroundImageUrl = (activeGame?.useAlternativeBackground && activeGame?.alternativeBannerUrl)
-    ? activeGame.alternativeBannerUrl
-    : activeGame?.heroUrl || activeGame?.bannerUrl || activeGame?.boxArtUrl || '';
-
-  // Check if this is an Alpha build
-  const isAlphaBuild = __BUILD_PROFILE__ === 'alpha' || (typeof process !== 'undefined' && process.env?.NODE_ENV === 'development');
-
   return (
-    <div className="h-screen bg-[radial-gradient(ellipse_at_top,_var(--tw-gradient-stops))] from-slate-900 via-[#0f172a] to-black text-white flex flex-col overflow-hidden relative">
-      {/* Background - Image or Color */}
-      {backgroundMode === 'image' && backgroundImageUrl ? (
-        <div
-          key={backgroundImageUrl}
-          className="fixed inset-0 pointer-events-none"
-          style={{
-            backgroundImage: `url(${backgroundImageUrl})`,
-            backgroundSize: 'cover',
-            backgroundPosition: 'center',
-            backgroundRepeat: 'no-repeat',
-            filter: `blur(${backgroundBlur}px) brightness(${currentBackgroundBrightness})`,
-            transform: backgroundBlur > 0 ? `scale(${1 + (backgroundBlur * 0.002)})` : 'none',
-            zIndex: 0,
-            transition: 'opacity 600ms ease-in-out',
-            animation: 'fadeIn 600ms ease-in-out',
-          }}
-        />
-      ) : (
-        <div
-          className="fixed inset-0 pointer-events-none"
-          style={{
-            backgroundColor: backgroundColor,
-            zIndex: 0,
-          }}
-        />
-      )}
-
-      {/* Aurora glow effect behind the top area */}
-      <div className="absolute top-0 left-0 right-0 h-96 bg-blue-500/10 blur-[100px] pointer-events-none" style={{ zIndex: 1 }} />
-
-      {/* Content wrapper with proper z-index */}
-      <div className="relative z-10 flex flex-col h-full">
-        {/* Menu Bar - Fixed at top */}
-        <MenuBar
-          onScanFolder={handleScanFolder}
-          onUpdateSteamLibrary={handleUpdateSteamLibrary}
-          onUpdateLibrary={handleUpdateSteamLibrary}
-          onGameManager={() => setIsGameManagerOpen(true)}
-          onConfigureSteam={() => setIsSteamConfigOpen(true)}
-          onOnyxSettings={() => {
-            setOnyxSettingsInitialTab('general');
-            setIsOnyxSettingsOpen(true);
-          }}
-          onAPISettings={() => {
-            setOnyxSettingsInitialTab('apis');
-            setIsOnyxSettingsOpen(true);
-          }}
-          onAbout={() => {
-            setOnyxSettingsInitialTab('about');
-            setIsOnyxSettingsOpen(true);
-          }}
-          onShowLibraryTutorial={() => setShowLibraryTutorial(true)}
-          onExit={handleExit}
-          onBugReport={isAlphaBuild ? () => setIsBugReportOpen(true) : undefined}
-          searchQuery={searchQuery}
-          onSearchChange={setSearchQuery}
-          selectedCategory={selectedCategory}
-          onCategoryChange={setSelectedCategory}
-          allCategories={allCategories}
-          categoryCounts={categoryCounts}
-          pinnedCategories={pinnedCategories}
-          onTogglePinCategory={handleTogglePinCategory}
-          onReorderPinnedCategories={setPinnedCategories}
-          sortBy={sortBy}
-          onSortChange={setSortBy}
-          hasFavoriteGames={hasFavoriteGames}
-          hasVRCategory={hasVRCategory}
-          hasAppsCategory={hasAppsCategory}
-          hasHiddenGames={hasHiddenGames}
-          hideVRTitles={hideVRTitles}
-          hideAppsTitles={hideAppsTitles}
-          showCategoriesInGameList={showCategoriesByView[viewMode] ?? false}
-          onToggleHideVRTitles={() => setHideVRTitles(prev => !prev)}
-          onToggleHideAppsTitles={() => setHideAppsTitles(prev => !prev)}
-          launchers={allLaunchers}
-          selectedLauncher={selectedLauncher}
-          onLauncherChange={setSelectedLauncher}
-          topBarPositions={topBarPositions}
-          onTopBarPositionsChange={async (positions) => {
-            setTopBarPositions(positions);
-            try {
-              await window.electronAPI.savePreferences({ topBarPositions: positions });
-            } catch (error) {
-              console.error('Error saving top bar positions:', error);
-            }
-          }}
-        />
-
-        {/* Top Bar - Hidden by default, shown when menu is open */}
-        {showTopBar && (
-          <TopBar
-            onSearch={setSearchQuery}
-            onRefresh={loadLibrary}
-            onFolder={() => handleScanFolder()}
-            onGridToggle={() => setViewMode(viewMode === 'grid' ? 'list' : 'grid')}
-            onSettings={() => setIsSteamConfigOpen(true)}
-            viewMode={viewMode}
-            notificationCount={0}
-          />
-        )}
-
-        {/* Main Content Area */}
-        <div className={`flex-1 flex overflow-hidden relative pt-10 ${isViewFlippedByView[viewMode] ? 'flex-row-reverse' : ''}`}>
-          {/* Left Panel - Game Library (flexible width, full width in carousel/coverflow mode) */}
-          <div className={`flex flex-col overflow-hidden ${viewMode === 'carousel' || viewMode === 'coverflow' ? 'w-full' : 'flex-1'}`}>
-            {/* Game Grid */}
-            <div
-              ref={gridContainerRef}
-              className={`flex-1 overflow-y-auto relative z-10 ${viewMode === 'carousel' || viewMode === 'coverflow' ? '' : (showCategoriesByView[viewMode] && (viewMode === 'grid' || viewMode === 'list' || viewMode === 'logo') ? 'p-0' : 'p-4')}`}
-              onContextMenuCapture={(e) => {
-                // Capture-phase fallback: open menu on any non-card area before children stop propagation
-                const target = e.target as HTMLElement;
-                if (!target.closest('[data-game-card]')) {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  setGameContextMenu(null);
-                  setRightClickMenu({ x: e.clientX, y: e.clientY });
-                }
-              }}
-              onContextMenu={(e) => {
-                const target = e.target as HTMLElement;
-                if (!target.closest('[data-game-card]')) {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  setGameContextMenu(null);
-                  setRightClickMenu({ x: e.clientX, y: e.clientY });
-                }
-              }}
-            >
-              {loading && (
-                <div className="text-center py-8">
-                  <p className="text-gray-100">Loading game library...</p>
-                </div>
-              )}
-
-              {error && (
-                <div className="bg-red-900/20 border border-red-500 rounded p-4 mb-4">
-                  <p className="text-red-300">Error: {error}</p>
-                </div>
-              )}
-
-              {!loading && !error && (
-                <div className="h-full flex flex-col">
-                  {showCategoriesByView[viewMode] && viewMode !== 'carousel' && viewMode !== 'coverflow' && pinnedCategories.length > 0 && (categoriesPositionByView[viewMode] ?? 'top') === 'top' && (
-                    <div
-                      className={`flex items-center gap-2 px-6 py-4 overflow-x-auto no-scrollbar ${(categoriesAlignmentByView[viewMode] ?? 'left') === 'center' ? 'justify-center' : (categoriesAlignmentByView[viewMode] ?? 'left') === 'right' ? 'justify-end' : 'justify-start'
-                        }`}
-                      style={{ WebkitAppRegion: 'no-drag' } as React.CSSProperties}
-                    >
-                      <button
-                        onClick={() => setSelectedCategory(null)}
-                        style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                        className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === null
-                          ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                          : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                          }`}
-                      >
-                        All Games
-                      </button>
-                      {hasFavoriteGames && (
-                        <button
-                          onClick={() => setSelectedCategory(selectedCategory === 'favorites' ? null : 'favorites')}
-                          style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                          className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === 'favorites'
-                            ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                            : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                            }`}
-                        >
-                          Favorites
-                        </button>
-                      )}
-                      {pinnedCategories.map((cat) => (
-                        <button
-                          key={cat}
-                          onClick={() => setSelectedCategory(selectedCategory === cat ? null : cat)}
-                          style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                          className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === cat
-                            ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                            : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                            }`}
-                        >
-                          {cat}
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                  {filteredGames.length > 0 ? (
-                    <div className={`flex-1 overflow-y-auto animate-onyx-grid-fade ${showCategoriesByView[viewMode] && viewMode !== 'carousel' && viewMode !== 'coverflow' ? ((categoriesPositionByView[viewMode] ?? 'top') === 'top' ? 'px-4 pb-4 pt-0' : 'px-4 pt-4 pb-0') : ''}`}>
-                      {viewMode === 'grid' || viewMode === 'logo' ? (
-                        <LibraryGrid
-                          games={filteredGames}
-                          onReorder={handleReorder}
-                          onPlay={handlePlay}
-                          onGameClick={handleGameClick}
-                          onEdit={handleEditGame}
-                          onEditImages={handleEditImages}
-                          onEditCategories={handleEditCategories}
-                          onFavorite={handleToggleFavorite}
-                          onPin={handleTogglePin}
-                          onFixMatch={handleFixMatch}
-                          onHide={handleHideGame}
-                          onUnhide={handleUnhideGame}
-                          isHiddenView={selectedCategory === 'hidden'}
-                          gridSize={gridSize}
-                          logoSize={logoSize}
-                          onGridSizeChange={setGridSize}
-                          gameTilePadding={gameTilePadding}
-                          hideGameTitles={hideGameTitles}
-                          showLogoOverBoxart={showLogoOverBoxart}
-                          logoPosition={logoPosition}
-                          useLogosInsteadOfBoxart={viewMode === 'logo'}
-                          autoSizeToFit={autoSizeToFit}
-                          logoBackgroundColor={logoBackgroundColor}
-                          logoBackgroundOpacity={logoBackgroundOpacity}
-                          descriptionSize={gridDescriptionSize}
-                          onGameContextMenu={(game: Game, x: number, y: number) => {
-                            setRightClickMenu(null);
-                            setGameContextMenu({ game, x, y });
-                          }}
-                          onEmptySpaceClick={(x: number, y: number) => {
-                            setGameContextMenu(null);
-                            setRightClickMenu({ x, y });
-                          }}
-                          viewMode={viewMode as 'grid' | 'logo'}
-                        />
-                      ) : viewMode === 'coverflow' ? (
-                        <LibraryCoverFlow
-                          games={filteredGames}
-                          onPlay={handlePlay}
-                          onGameClick={handleGameClick}
-                          onEdit={handleEditGame}
-                          onEditImages={handleEditImages}
-                          onEditCategories={handleEditCategories}
-                          onFavorite={handleToggleFavorite}
-                          onPin={handleTogglePin}
-                          onFixMatch={handleFixMatch}
-                          onHide={handleHideGame}
-                          onUnhide={handleUnhideGame}
-                          isHiddenView={selectedCategory === 'hidden'}
-                          activeGameId={activeGameId}
-                          coverSize={coverFlowCoverSize}
-                          reflectionStrength={coverFlowReflection / 100}
-                          verticalOffset={coverFlowVerticalOffset}
-                          sideOpacity={coverFlowSideOpacity}
-                          showButtons={coverFlowShowButtons}
-                          buttonPosition={coverFlowButtonPosition}
-                          buttonColors={coverFlowButtonColors}
-                          onEmptySpaceRightClick={(x, y) => {
-                            setGameContextMenu(null);
-                            setRightClickMenu({ x, y });
-                          }}
-                        />
-                      ) : viewMode === 'carousel' ? (
-                        <LibraryCarousel
-                          games={filteredGames}
-                          onPlay={handlePlay}
-                          onGameClick={handleGameClick}
-                          onEdit={handleEditGame}
-                          onEditImages={handleEditImages}
-                          onEditCategories={handleEditCategories}
-                          onFavorite={handleToggleFavorite}
-                          onPin={handleTogglePin}
-                          onFixMatch={handleFixMatch}
-                          onHide={handleHideGame}
-                          onUnhide={handleUnhideGame}
-                          isHiddenView={selectedCategory === 'hidden'}
-                          activeGameId={activeGameId}
-                          selectedBoxArtSize={selectedBoxArtSize}
-                          gameTilePadding={carouselGameTilePadding}
-                          showCarouselDetails={showCarouselDetails}
-                          showCarouselLogos={showCarouselLogos}
-                          detailsBarSize={detailsBarSize}
-                          onDetailsBarSizeChange={(size) => {
-                            setDetailsBarSize(size);
-                            window.electronAPI.savePreferences({ detailsBarSize: size });
-                          }}
-                          carouselLogoSize={carouselLogoSize}
-                          onCarouselLogoSizeChange={(size) => {
-                            setCarouselLogoSize(size);
-                            window.electronAPI.savePreferences({ carouselLogoSize: size });
-                          }}
-                          carouselButtonSize={carouselButtonSize}
-                          onCarouselButtonSizeChange={(size) => {
-                            setCarouselButtonSize(size);
-                            window.electronAPI.savePreferences({ carouselButtonSize: size });
-                          }}
-                          carouselDescriptionSize={carouselDescriptionSize}
-                          onCarouselDescriptionSizeChange={(size) => {
-                            setCarouselDescriptionSize(size);
-                            window.electronAPI.savePreferences({ carouselDescriptionSize: size });
-                          }}
-                          onEmptySpaceRightClick={(x, y) => {
-                            setGameContextMenu(null);
-                            setRightClickMenu({ x, y });
-                          }}
-                          isViewFlipped={isViewFlippedByView[viewMode]}
-                          carouselButtonColors={carouselButtonColors}
-                        />
-                      ) : (
-                        <LibraryListView
-                          games={filteredGames}
-                          onPlay={handlePlay}
-                          onGameClick={handleGameClick}
-                          onEdit={handleEditGame}
-                          onEditImages={handleEditImages}
-                          onEditCategories={handleEditCategories}
-                          onFavorite={handleToggleFavorite}
-                          onPin={handleTogglePin}
-                          onFixMatch={handleFixMatch}
-                          onHide={handleHideGame}
-                          onUnhide={handleUnhideGame}
-                          isHiddenView={selectedCategory === 'hidden'}
-                          hideGameTitles={hideGameTitles}
-                          listViewOptions={listViewOptions}
-                          listViewSize={listViewSize}
-                          onEmptySpaceClick={(x, y) => {
-                            setGameContextMenu(null);
-                            setRightClickMenu({ x, y });
-                          }}
-                        />
-                      )}
-                    </div>
-                  ) : (
-                    <WelcomeScreen
-                      onScanGames={() => {
-                        setAutoStartScan(true);
-                        setIsImportWorkbenchOpen(true);
-                      }}
-                      onAddFolder={handleAddFolder}
-                      onOpenSettings={() => setIsAPISettingsOpen(true)}
-                    />
-                  )}
-                  {showCategoriesByView[viewMode] && viewMode !== 'carousel' && viewMode !== 'coverflow' && pinnedCategories.length > 0 && (categoriesPositionByView[viewMode] ?? 'top') === 'bottom' && (
-                    <div
-                      className={`flex items-center gap-2 px-6 py-4 overflow-x-auto no-scrollbar ${(categoriesAlignmentByView[viewMode] ?? 'left') === 'center' ? 'justify-center' : (categoriesAlignmentByView[viewMode] ?? 'left') === 'right' ? 'justify-end' : 'justify-start'
-                        }`}
-                      style={{ WebkitAppRegion: 'no-drag' } as React.CSSProperties}
-                    >
-                      <button
-                        onClick={() => setSelectedCategory(null)}
-                        style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                        className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === null
-                          ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                          : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                          }`}
-                      >
-                        All Games
-                      </button>
-                      {hasFavoriteGames && (
-                        <button
-                          onClick={() => setSelectedCategory(selectedCategory === 'favorites' ? null : 'favorites')}
-                          style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                          className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === 'favorites'
-                            ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                            : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                            }`}
-                        >
-                          Favorites
-                        </button>
-                      )}
-                      {pinnedCategories.map((cat) => (
-                        <button
-                          key={cat}
-                          onClick={() => setSelectedCategory(selectedCategory === cat ? null : cat)}
-                          style={{ fontSize: `${categoriesSizeByView[viewMode] ?? 12}px` }}
-                          className={`px-3 py-1.5 rounded-full font-medium transition-all whitespace-nowrap ${selectedCategory === cat
-                            ? 'bg-blue-600/40 text-blue-100 border border-blue-500/40 shadow-sm shadow-blue-500/20'
-                            : 'bg-gray-800/40 text-gray-400 hover:bg-gray-700/60 hover:text-gray-200 border border-gray-700/20'
-                            }`}
-                        >
-                          {cat}
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Right Panel - Game Details (hidden in carousel/coverflow mode and when no games exist) */}
-          {viewMode !== 'carousel' && viewMode !== 'coverflow' && filteredGames.length > 0 && (
-            <GameDetailsPanel
-              game={activeGame}
-              isLaunching={launchingGameId === activeGame?.id}
-              isRunning={activeGame ? runningGames.has(activeGame.id) : false}
-              onPlay={handlePlay}
-              onSaveGame={handleSaveGame}
-              onUpdateGameInState={updateGameInState}
-              viewMode={viewMode}
-              onOpenInGameManager={(game, tab) => {
-                setGameManagerInitialGameId(game.id);
-                setGameManagerInitialTab(tab);
-                setIsGameManagerOpen(true);
-              }}
-              onFavorite={handleToggleFavorite}
-              onEdit={handleEditGame}
-              onEditImages={handleEditImages}
-              onEditCategories={handleEditCategories}
-              onPin={handleTogglePin}
-              onFixMatch={handleFixMatch}
-              onHide={handleHideGame}
-              onUnhide={handleUnhideGame}
-              isHiddenView={selectedCategory === 'hidden'}
-              onRightClick={(x, y) => {
-                setGameContextMenu(null);
-                setRightClickMenu({ x, y });
-              }}
-              panelWidth={currentPanelWidth}
-              onPanelWidthChange={(width) => {
-                setPanelWidth(width);
-                const newByView = { ...panelWidthByViewState, [viewMode]: width };
-                setPanelWidthByViewState(newByView);
-                window.electronAPI.savePreferences({ panelWidthByView: newByView });
-              }}
-              rightPanelLogoSize={rightPanelLogoSize}
-              rightPanelBoxartPosition={rightPanelBoxartPosition}
-              rightPanelBoxartSize={rightPanelBoxartSize}
-              rightPanelTextSize={rightPanelTextSize}
-              rightPanelButtonSize={rightPanelButtonSize}
-              rightPanelButtonLocation={rightPanelButtonLocation}
-              detailsPanelOpacity={detailsPanelOpacity}
-              fanartHeight={currentFanartHeight}
-              onFanartHeightChange={(height) => {
-                const newByView = { ...fanartHeightByView, [viewMode]: height };
-                setFanartHeightByView(newByView);
-                window.electronAPI.savePreferences({ fanartHeightByView: newByView });
-              }}
-              descriptionWidth={currentDescriptionWidth}
-              onDescriptionWidthChange={(width) => {
-                const newByView = { ...descriptionWidthByView, [viewMode]: width };
-                setDescriptionWidthByView(newByView);
-                window.electronAPI.savePreferences({ descriptionWidthByView: newByView });
-              }}
-              isViewFlipped={isViewFlippedByView[viewMode]}
-              rightPanelButtonColors={
-                viewMode === 'grid' ? gridButtonColors :
-                  viewMode === 'list' ? listButtonColors :
-                    viewMode === 'logo' ? logoButtonColors :
-                      rightPanelButtonColors
-              }
-              linkDisplayOrder={linkDisplayOrder}
-              visibleLinkTypes={visibleLinkTypes}
-            />
-          )}
-        </div>
-      </div>
-
-      {/* Add Game Modal */}
-      <AddGameModal
-        isOpen={isModalOpen}
-        onClose={() => setIsModalOpen(false)}
-        onAdd={handleAddGame}
-      />
-
-
-      {/* Game Metadata Editor */}
-      {selectedExecutable && (
-        <GameMetadataEditor
-          isOpen={isMetadataEditorOpen}
-          onClose={() => {
-            setIsMetadataEditorOpen(false);
-            setSelectedExecutable(null);
-          }}
-          executable={selectedExecutable}
-          onSave={handleSaveGameWithMetadata}
-        />
-      )}
-
-      {/* Steam Configuration Modal */}
-      <SteamConfigModal
-        isOpen={isSteamConfigOpen}
-        onClose={() => setIsSteamConfigOpen(false)}
-        onScan={handleSteamConfigScan}
-      />
-
-
-      {/* Categories Editor Modal */}
-      <CategoriesEditor
-        isOpen={isCategoriesEditorOpen}
-        game={editingCategoriesGame}
-        onClose={() => {
-          setIsCategoriesEditorOpen(false);
-          setEditingCategoriesGame(null);
-        }}
-        onSave={async (game) => {
-          await handleSaveGame(game);
-        }}
-        allCategories={allCategories}
-      />
-
-
-      {/* Onyx Settings Modal */}
-      <OnyxSettingsModal
-        isOpen={isOnyxSettingsOpen}
-        onClose={() => setIsOnyxSettingsOpen(false)}
-        initialTab={onyxSettingsInitialTab}
-        onShowImportModal={(games, appType) => {
-          // Use ImportWorkbench instead of SteamImportModal
-          setScannedSteamGames(games);
-          setImportAppType(appType || 'steam');
-          setIsImportWorkbenchOpen(true);
-        }}
-        onSave={async () => {
-          // Reload preferences after saving to update UI immediately
-          try {
-            const prefs = await window.electronAPI.getPreferences();
-            if (prefs.hideGameTitles !== undefined) {
-              setHideGameTitles(prefs.hideGameTitles);
-            }
-            if (prefs.gameTilePadding !== undefined) {
-              setGameTilePadding(prefs.gameTilePadding);
-            }
-            if (prefs.showLogoOverBoxart !== undefined) {
-              setShowLogoOverBoxart(prefs.showLogoOverBoxart);
-            }
-            if (prefs.logoPosition !== undefined) {
-              setLogoPosition(prefs.logoPosition);
-            }
-            if (prefs.showLogoOverBoxart !== undefined) {
-              setShowLogoOverBoxart(prefs.showLogoOverBoxart);
-            }
-            if (prefs.logoPosition !== undefined) {
-              setLogoPosition(prefs.logoPosition);
-            }
-            // Also update other settings that might have changed
-            if (prefs.minimizeToTray !== undefined) {
-              // These are handled by the modal, but we can reload them too if needed
-            }
-
-            if (prefs.confirmGameLaunch !== undefined) {
-              setConfirmGameLaunch(prefs.confirmGameLaunch);
-            }
-            if (prefs.linkDisplayOrder && prefs.linkDisplayOrder.length > 0) {
-              setLinkDisplayOrder(prefs.linkDisplayOrder);
-            }
-            if (prefs.visibleLinkTypes && Object.keys(prefs.visibleLinkTypes).length > 0) {
-              setVisibleLinkTypes(prefs.visibleLinkTypes);
-            }
-
-            // Reload library if app configs were saved
-            await loadLibrary();
-          } catch (error) {
-            console.error('Error reloading preferences after save:', error);
-          }
-        }}
-      />
-
-      {/* API Settings Modal */}
-      <APISettingsModal
-        isOpen={isAPISettingsOpen}
-        onClose={() => setIsAPISettingsOpen(false)}
-      />
-
-
-      {/* Update Library Modal */}
-      <UpdateLibraryModal
-        isOpen={isUpdateLibraryOpen}
-        onClose={() => setIsUpdateLibraryOpen(false)}
-        onUpdate={() => {
-          loadLibrary();
-        }}
-        onShowImportModal={(games, appType = 'steam') => {
-          // Use ImportWorkbench instead of SteamImportModal
-          setScannedSteamGames(games);
-          setImportAppType(appType);
-          setIsImportWorkbenchOpen(true);
-        }}
-      />
-
-      {/* Game Importer */}
-      <ImportWorkbench
-        isOpen={isImportWorkbenchOpen}
-        autoStartScan={autoStartScan}
-        onClose={() => {
-          setIsImportWorkbenchOpen(false);
-          setAutoStartScan(false);
-          setImportWorkbenchFolderPath(undefined);
-          setScannedSteamGames([]);
-          setImportAppType('steam');
-        }}
-        existingLibrary={games}
-        preScannedGames={scannedSteamGames && scannedSteamGames.length > 0 ? scannedSteamGames : undefined}
-        onImport={async (games, onProgress) => {
-          try {
-            for (let i = 0; i < games.length; i++) {
-              const game = games[i];
-              onProgress?.(i + 1, games.length, 'Saving games', `Saving ${game.title}...`);
-              await window.electronAPI.saveGame(game);
-            }
-
-            // Only fetch artwork for games that don't already have it (e.g. already fetched in importer)
-            const gamesNeedingArtwork = games.filter(
-              (g) => !g.bannerUrl || !g.iconUrl || !g.alternativeBannerUrl
-            );
-            const BATCH_SIZE = 5;
-            if (gamesNeedingArtwork.length > 0) {
-              for (let i = 0; i < gamesNeedingArtwork.length; i += BATCH_SIZE) {
-                const batch = gamesNeedingArtwork.slice(i, i + BATCH_SIZE);
-                const firstTitle = batch[0]?.title ?? '';
-                onProgress?.(i, gamesNeedingArtwork.length, 'Fetching artwork', firstTitle);
-                console.log(`[Import] Processing batch ${i / BATCH_SIZE + 1} (${batch.length} games)`);
-
-                await Promise.all(batch.map(async (game) => {
-                  try {
-                    console.log(`[Import] Fetching banners for: ${game.title}`);
-
-                    const metadata = await window.electronAPI.searchArtwork(game.title, (game as any).appId);
-
-                    if (metadata) {
-                      let updatedGame = { ...game };
-                      let updated = false;
-
-                      if (metadata.bannerUrl && !game.bannerUrl) {
-                        updatedGame.bannerUrl = metadata.bannerUrl;
-                        updated = true;
-                      }
-                      if (metadata.alternativeBannerUrl && !game.alternativeBannerUrl) {
-                        updatedGame.alternativeBannerUrl = metadata.alternativeBannerUrl;
-                        updated = true;
-                      }
-                      if (metadata.iconUrl && !game.iconUrl) {
-                        updatedGame.iconUrl = metadata.iconUrl;
-                        updated = true;
-                      }
-
-                      try {
-                        const steamAppId = (game as any).appId;
-                        const bannerSearch = await window.electronAPI.searchImages(game.title, 'banner', steamAppId);
-
-                        if (bannerSearch?.success && bannerSearch.images) {
-                          const allBannerUrls: string[] = [];
-
-                          if (Array.isArray(bannerSearch.images)) {
-                            bannerSearch.images.forEach((item: any) => {
-                              if (item.images && Array.isArray(item.images)) {
-                                item.images.forEach((img: any) => {
-                                  const url = img.url || img.bannerUrl;
-                                  if (url && !allBannerUrls.includes(url)) allBannerUrls.push(url);
-                                });
-                              } else if (item.url || item.bannerUrl) {
-                                const url = item.url || item.bannerUrl;
-                                if (url && !allBannerUrls.includes(url)) allBannerUrls.push(url);
-                              }
-                            });
-                          }
-
-                          if (!updatedGame.bannerUrl && allBannerUrls.length > 0) {
-                            updatedGame.bannerUrl = allBannerUrls[0];
-                            updated = true;
-                          }
-                          if (allBannerUrls.length > 1) {
-                            const altUrl = allBannerUrls.find(url => url !== updatedGame.bannerUrl) || allBannerUrls[1];
-                            if (altUrl && altUrl !== updatedGame.bannerUrl) {
-                              updatedGame.alternativeBannerUrl = altUrl;
-                              updated = true;
-                            }
-                          }
-                        }
-                      } catch (searchErr) {
-                        console.error(`[Import] Banner search error for ${game.title}:`, searchErr);
-                      }
-
-                      if (updated) {
-                        await window.electronAPI.saveGame(updatedGame);
-                      }
-                    }
-                  } catch (err) {
-                    console.error(`[Import] Failed to fetch metadata for ${game.title}:`, err);
-                  }
-                }));
-                onProgress?.(i + batch.length, gamesNeedingArtwork.length, 'Fetching artwork', batch[batch.length - 1]?.title ?? '');
-              }
-            }
-
-            onProgress?.(games.length, games.length, 'Finishing', 'Reloading library...');
-            await loadLibrary();
-            showToast(`Successfully imported ${games.length} ${games.length === 1 ? 'game' : 'games'}`, 'success');
-            setIsImportWorkbenchOpen(false);
-            setAutoStartScan(false);
-            setImportWorkbenchFolderPath(undefined);
-            setScannedSteamGames([]);
-            setImportAppType('steam');
-            const prefs = await window.electronAPI.getPreferences();
-            if (!prefs.hasSeenPostImportTutorial) {
-              setShowLibraryTutorial(true);
-              await window.electronAPI.savePreferences({ hasSeenPostImportTutorial: true });
-            }
-          } catch (err) {
-            console.error('Error importing games:', err);
-            showToast('Failed to import games', 'error');
-          }
-        }}
-      />
-
-      {/* Game Manager */}
-      <GameManager
-        isOpen={isGameManagerOpen}
-        onClose={() => {
-          setIsGameManagerOpen(false);
-          setGameManagerInitialGameId(null);
-          setGameManagerInitialTab('images');
-        }}
-        games={games}
-        initialGameId={gameManagerInitialGameId}
-        initialTab={gameManagerInitialTab}
-        onSaveGame={async (game, oldGame) => {
-          // Get old game if not provided
-          if (!oldGame) {
-            oldGame = games.find(g => g.id === game.id);
-          }
-          await window.electronAPI.saveGame(game, oldGame);
-          // Check if it's just an image update
-          const isImageUpdate = oldGame && (
-            game.boxArtUrl !== oldGame.boxArtUrl ||
-            game.bannerUrl !== oldGame.bannerUrl ||
-            game.alternativeBannerUrl !== oldGame.alternativeBannerUrl ||
-            game.logoUrl !== oldGame.logoUrl ||
-            game.heroUrl !== oldGame.heroUrl ||
-            game.iconUrl !== oldGame.iconUrl
-          );
-
-          if (isImageUpdate) {
-            // Update the game in state without reloading - this updates the main app immediately
-            updateGameInState(game);
-          } else {
-            // For non-image updates, reload the library
-            await loadLibrary();
-          }
-        }}
-        onDeleteGame={async (gameId) => {
-          await deleteGame(gameId);
-          await loadLibrary();
-        }}
-        onReloadLibrary={loadLibrary}
-      />
-
-      {/* Right Click Menu */}
-      {rightClickMenu && (
-        <RightClickMenu
-          x={rightClickMenu.x}
-          y={rightClickMenu.y}
-          onClose={() => setRightClickMenu(null)}
-          viewMode={viewMode}
-          onViewModeChange={setViewMode}
-          activeGame={activeGame || undefined}
-          onActiveGameChange={(game) => {
-            setActiveGameId(game.id);
-            updateGameInState(game);
-            // Note: Do NOT save here - let RightClickMenu handle saves with debouncing
-            // This prevents duplicate saves and flickering on logo size changes
-          }}
-          gridSize={gridSize}
-          onGridSizeChange={setGridSize}
-          logoSize={logoSize}
-          onLogoSizeChange={setLogoSize}
-          listSize={listViewSize}
-          onListSizeChange={setListViewSize}
-          listViewOptions={listViewOptions}
-          onListViewOptionsChange={(options) => {
-            setListViewOptions({
-              ...defaultListViewOptions,
-              ...options,
-              showLauncher: options.showLauncher ?? true,
-              showLogos: options.showLogos ?? false,
-            });
-            window.electronAPI.savePreferences({ listViewOptions: options });
-          }}
-          gameTilePadding={gameTilePadding}
-          onGameTilePaddingChange={setGameTilePadding}
-          backgroundBlur={backgroundBlur}
-          onBackgroundBlurChange={setBackgroundBlur}
-          backgroundBrightness={currentBackgroundBrightness}
-          onBackgroundBrightnessChange={(brightness: number) => {
-            const newByView = { ...backgroundBrightnessByView, [viewMode]: brightness };
-            setBackgroundBrightnessByView(newByView);
-            window.electronAPI.savePreferences({ backgroundBrightnessByView: newByView });
-          }}
-          selectedBoxArtSize={selectedBoxArtSize}
-          onSelectedBoxArtSizeChange={setSelectedBoxArtSize}
-          panelWidth={currentPanelWidth}
-          onPanelWidthChange={(width) => {
-            setPanelWidth(width);
-            const newByView = { ...panelWidthByViewState, [viewMode]: width };
-            setPanelWidthByViewState(newByView);
-            window.electronAPI.savePreferences({ panelWidthByView: newByView });
-          }}
-          carouselLogoSize={carouselLogoSize}
-          onCarouselLogoSizeChange={(size) => {
-            setCarouselLogoSize(size);
-            window.electronAPI.savePreferences({ carouselLogoSize: size });
-          }}
-          detailsBarSize={detailsBarSize}
-          onDetailsBarSizeChange={(size) => {
-            setDetailsBarSize(size);
-            window.electronAPI.savePreferences({ detailsBarSize: size });
-          }}
-          showCarouselDetails={showCarouselDetails}
-          onShowCarouselDetailsChange={(show) => {
-            setShowCarouselDetails(show);
-            window.electronAPI.savePreferences({ showCarouselDetails: show });
-          }}
-          showCarouselLogos={showCarouselLogos}
-          onShowCarouselLogosChange={(show) => {
-            setShowCarouselLogos(show);
-            window.electronAPI.savePreferences({ showCarouselLogos: show });
-          }}
-          carouselButtonSize={carouselButtonSize}
-          onCarouselButtonSizeChange={(size) => {
-            setCarouselButtonSize(size);
-            window.electronAPI.savePreferences({ carouselButtonSize: size });
-          }}
-          carouselDescriptionSize={carouselDescriptionSize}
-          onCarouselDescriptionSizeChange={(size) => {
-            setCarouselDescriptionSize(size);
-            window.electronAPI.savePreferences({ carouselDescriptionSize: size });
-          }}
-          showCategoriesInGameList={showCategoriesByView[viewMode] ?? false}
-          onShowCategoriesInGameListChange={(show) => {
-            const newByView = { ...showCategoriesByView, [viewMode]: show };
-            setShowCategoriesByView(newByView);
-            window.electronAPI.savePreferences({ showCategoriesInGameListByView: newByView });
-          }}
-          categoriesPosition={categoriesPositionByView[viewMode] ?? 'top'}
-          onCategoriesPositionChange={(position: 'top' | 'bottom') => {
-            const newByView = { ...categoriesPositionByView, [viewMode]: position };
-            setCategoriesPositionByView(newByView);
-            window.electronAPI.savePreferences({ categoriesPositionByView: newByView });
-          }}
-          categoriesTopAlignment={categoriesAlignmentByView[viewMode] ?? 'left'}
-          onCategoriesTopAlignmentChange={(alignment: 'left' | 'center' | 'right') => {
-            const newByView = { ...categoriesAlignmentByView, [viewMode]: alignment };
-            setCategoriesAlignmentByView(newByView);
-            window.electronAPI.savePreferences({ categoriesAlignmentByView: newByView });
-          }}
-          categoriesTopSize={categoriesSizeByView[viewMode] ?? 12}
-          onCategoriesTopSizeChange={(size: number) => {
-            const newByView = { ...categoriesSizeByView, [viewMode]: size };
-            setCategoriesSizeByView(newByView);
-            window.electronAPI.savePreferences({ categoriesSizeByView: newByView });
-          }}
-          showLogoOverBoxart={showLogoOverBoxart}
-          onShowLogoOverBoxartChange={(show) => {
-            setShowLogoOverBoxart(show);
-            window.electronAPI.savePreferences({ showLogoOverBoxart: show });
-          }}
-          logoPosition={logoPosition}
-          onLogoPositionChange={(position) => {
-            setLogoPosition(position);
-            window.electronAPI.savePreferences({ logoPosition: position });
-          }}
-          logoBackgroundColor={logoBackgroundColor}
-          onLogoBackgroundColorChange={(color: string) => {
-            setLogoBackgroundColor(color);
-            window.electronAPI.savePreferences({ logoBackgroundColor: color });
-          }}
-          logoBackgroundOpacity={logoBackgroundOpacity}
-          onLogoBackgroundOpacityChange={(opacity: number) => {
-            setLogoBackgroundOpacity(opacity);
-            window.electronAPI.savePreferences({ logoBackgroundOpacity: opacity });
-          }}
-          rightPanelLogoSize={rightPanelLogoSize}
-          onRightPanelLogoSizeChange={(size) => {
-            setRightPanelLogoSize(size);
-            window.electronAPI.savePreferences({ rightPanelLogoSize: size });
-          }}
-          rightPanelBoxartPosition={rightPanelBoxartPosition}
-          onRightPanelBoxartPositionChange={(position) => {
-            setRightPanelBoxartPosition(position);
-            window.electronAPI.savePreferences({ rightPanelBoxartPosition: position });
-          }}
-          rightPanelBoxartSize={rightPanelBoxartSize}
-          onRightPanelBoxartSizeChange={(size) => {
-            setRightPanelBoxartSize(size);
-            window.electronAPI.savePreferences({ rightPanelBoxartSize: size });
-          }}
-          rightPanelTextSize={rightPanelTextSize}
-          onRightPanelTextSizeChange={(size) => {
-            setRightPanelTextSize(size);
-            window.electronAPI.savePreferences({ rightPanelTextSize: size });
-          }}
-          rightPanelButtonSize={rightPanelButtonSize}
-          onRightPanelButtonSizeChange={(size) => {
-            setRightPanelButtonSize(size);
-            window.electronAPI.savePreferences({ rightPanelButtonSize: size });
-          }}
-          rightPanelButtonLocation={rightPanelButtonLocation}
-          onRightPanelButtonLocationChange={(location) => {
-            setRightPanelButtonLocation(location);
-            window.electronAPI.savePreferences({ rightPanelButtonLocation: location });
-          }}
-          isViewFlipped={isViewFlippedByView[viewMode]}
-          onViewFlipChange={(flipped) => {
-            const newByView = { ...isViewFlippedByView, [viewMode]: flipped };
-            setIsViewFlippedByView(newByView);
-            window.electronAPI.savePreferences({ isViewFlippedByView: newByView });
-          }}
-          detailsPanelOpacity={detailsPanelOpacity}
-          onDetailsPanelOpacityChange={(opacity) => {
-            setDetailsPanelOpacity(opacity);
-            window.electronAPI.savePreferences({ detailsPanelOpacity: opacity });
-          }}
-          fanartHeight={currentFanartHeight}
-          onFanartHeightChange={(height) => {
-            const newByView = { ...fanartHeightByView, [viewMode]: height };
-            setFanartHeightByView(newByView);
-            window.electronAPI.savePreferences({ fanartHeightByView: newByView });
-          }}
-          descriptionWidth={currentDescriptionWidth}
-          onDescriptionWidthChange={(width) => {
-            const newByView = { ...descriptionWidthByView, [viewMode]: width };
-            setDescriptionWidthByView(newByView);
-            window.electronAPI.savePreferences({ descriptionWidthByView: newByView });
-          }}
-          rightPanelButtonColors={rightPanelButtonColors}
-          onRightPanelButtonColorsChange={(colors) => {
-            setRightPanelButtonColors(colors);
-            window.electronAPI.savePreferences({ rightPanelButtonColors: colors });
-          }}
-          carouselButtonColors={carouselButtonColors}
-          onCarouselButtonColorsChange={(colors) => {
-            setCarouselButtonColors(colors);
-            window.electronAPI.savePreferences({ carouselButtonColors: colors });
-          }}
-          gridButtonColors={gridButtonColors}
-          onGridButtonColorsChange={(colors) => {
-            setGridButtonColors(colors);
-            window.electronAPI.savePreferences({ gridButtonColors: colors });
-          }}
-          listButtonColors={listButtonColors}
-          onListButtonColorsChange={(colors) => {
-            setListButtonColors(colors);
-            window.electronAPI.savePreferences({ listButtonColors: colors });
-          }}
-          logoButtonColors={logoButtonColors}
-          onLogoButtonColorsChange={(colors) => {
-            setLogoButtonColors(colors);
-            window.electronAPI.savePreferences({ logoButtonColors: colors });
-          }}
-          coverFlowCoverSize={coverFlowCoverSize}
-          onCoverFlowCoverSizeChange={(size) => {
-            setCoverFlowCoverSize(size);
-            window.electronAPI.savePreferences({ coverFlowCoverSize: size });
-          }}
-          coverFlowReflection={coverFlowReflection}
-          onCoverFlowReflectionChange={(value) => {
-            setCoverFlowReflection(value);
-            window.electronAPI.savePreferences({ coverFlowReflection: value });
-          }}
-          coverFlowVerticalOffset={coverFlowVerticalOffset}
-          onCoverFlowVerticalOffsetChange={(value: number) => {
-            setCoverFlowVerticalOffset(value);
-            window.electronAPI.savePreferences({ coverFlowVerticalOffset: value } as any);
-          }}
-          coverFlowSideOpacity={coverFlowSideOpacity}
-          onCoverFlowSideOpacityChange={(value: number) => {
-            setCoverFlowSideOpacity(value);
-            window.electronAPI.savePreferences({ coverFlowSideOpacity: value } as any);
-          }}
-          coverFlowShowButtons={coverFlowShowButtons}
-          onCoverFlowShowButtonsChange={(show) => {
-            setCoverFlowShowButtons(show);
-            window.electronAPI.savePreferences({ coverFlowShowButtons: show });
-          }}
-          coverFlowButtonPosition={coverFlowButtonPosition}
-          onCoverFlowButtonPositionChange={(pos) => {
-            setCoverFlowButtonPosition(pos);
-            window.electronAPI.savePreferences({ coverFlowButtonPosition: pos });
-          }}
-          coverFlowButtonColors={coverFlowButtonColors}
-          onCoverFlowButtonColorsChange={(colors) => {
-            setCoverFlowButtonColors(colors);
-            window.electronAPI.savePreferences({ coverFlowButtonColors: colors });
-          }}
-          onSettingsImported={refreshPreferences}
-        />
-      )}
-
-      {/* Game Context Menu */}
-      {gameContextMenu && (
-        <GameContextMenu
-          game={gameContextMenu.game}
-          x={gameContextMenu.x}
-          y={gameContextMenu.y}
-          onClose={() => setGameContextMenu(null)}
-          onPlay={handlePlay}
-          onEdit={handleEditGame}
-          onEditImages={handleEditImages}
-          onEditCategories={handleEditCategories}
-          onFavorite={handleToggleFavorite}
-          onPin={handleTogglePin}
-          onFixMatch={handleFixMatch}
-          onHide={handleHideGame}
-          onUnhide={handleUnhideGame}
-          isHiddenView={selectedCategory === 'hidden'}
-        />
-      )}
-
-      {/* Metadata Search Modal */}
-      {fixingGame && (
-        <MetadataSearchModal
-          isOpen={isMetadataSearchOpen}
-          onClose={() => {
-            setIsMetadataSearchOpen(false);
-            setFixingGame(null);
-          }}
-          game={fixingGame}
-          onSelect={handleSelectMetadataMatch}
-        />
-      )}
-
-      {/* Bug Report Modal */}
-      <BugReportModal
-        isOpen={isBugReportOpen}
-        onClose={() => setIsBugReportOpen(false)}
-      />
-
-      {/* Hide Game Confirmation Dialog */}
-      {hideConfirmation && (
-        <ConfirmationDialog
-          isOpen={true}
-          title={`Hide "${hideConfirmation.game.title}"?`}
-          message="This game will be hidden from your library view."
-          note="You can find hidden games by selecting the 'Hidden' category from the Categories dropdown."
-          confirmText="Hide"
-          cancelText="Cancel"
-          onConfirm={handleConfirmHide}
-          onCancel={handleCancelHide}
-        />
-      )}
-
-      {/* Launch Confirmation Dialog */}
-      {launchConfirmation && (
-        <ConfirmationDialog
-          isOpen={true}
-          title={`Launch "${launchConfirmation.game.title}"?`}
-          message="Are you sure you want to launch this game?"
-          confirmText="Launch"
-          cancelText="Cancel"
-          onConfirm={() => {
-            launchGame(launchConfirmation.game);
-            setLaunchConfirmation(null);
-          }}
-          onCancel={() => setLaunchConfirmation(null)}
-        />
-      )}
-
-      {/* Update Notification Modal - Persistent until user interacts */}
-      {updateNotification && (
-        <UpdateNotificationModal
-          isOpen={true}
-          version={updateNotification.version}
-          status={updateNotification.status}
-          error={updateNotification.error}
-          currentVersion={currentVersion}
-          changelogSource={changelogSource}
-          changelogLoading={changelogLoading}
-          changelogError={changelogError}
-          isTestMode={isUpdateModalTest}
-          onUpdateNow={async () => {
-            const result = await window.electronAPI.downloadUpdate?.();
-            if (!result?.success) {
-              setUpdateNotification(prev => prev ? { ...prev, status: 'error', error: result?.error ?? 'Download failed' } : null);
-            }
-          }}
-          onDismiss={() => {
-            setUpdateNotification(null);
-            setIsUpdateModalTest(false);
-            // Signal main process that update was dismissed (so startup scan can proceed)
-            if (!isUpdateModalTest) {
-              window.electronAPI.onUpdateDismissed?.();
-            }
-          }}
-          onInstall={() => {
-            window.electronAPI.quitAndInstall?.();
-          }}
-        />
-      )}
-
-      <LibraryTutorialModal
-        isOpen={showLibraryTutorial}
-        onClose={() => setShowLibraryTutorial(false)}
-        onOpenSettings={() => {
-          setOnyxSettingsInitialTab('general');
-          setIsOnyxSettingsOpen(true);
-        }}
-        onOpenUpdateLibrary={handleUpdateSteamLibrary}
-      />
-
-      {/* Toast Notification - app styling, slides up from bottom */}
-      {toast && (
-        <div className="fixed bottom-5 left-1/2 -translate-x-1/2 z-50 toast-slide-up">
-          <div
-            className={`px-5 py-3 rounded-xl shadow-2xl flex items-center gap-3 backdrop-blur-xl border max-w-[90vw] ${toast.type === 'success'
-              ? 'bg-slate-900/95 border-cyan-500/40 text-slate-100'
-              : 'bg-slate-900/95 border-red-500/40 text-slate-100'
-              }`}
-          >
-            <div className="flex-1 text-sm">{toast.message}</div>
-            <button
-              onClick={() => setToast(null)}
-              className="text-slate-400 hover:text-slate-100 transition-colors p-0.5 rounded"
-            >
-              <svg className="w-5 h-5 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-              </svg>
-            </button>
-          </div>
-        </div>
-      )}
-
-      {/* Missing Games Modal */}
-      {missingGames && missingGames.length > 0 && (
-        <MissingGamesModal
-          missingGames={missingGames}
-          onRemove={handleRemoveMissingGames}
-          onCancel={handleCancelMissingGames}
-        />
-      )}
-
-
-      {missingGames && (
-        <MissingGamesModal
-          missingGames={missingGames}
-          onRemove={async (gameIds) => {
-            await window.electronAPI.removeMissingGames(gameIds);
-            setMissingGames(null);
-            loadLibrary();
-          }}
-          onCancel={() => setMissingGames(null)}
-        />
-      )}
-
-
-
-      {/* Startup scan progress overlay */}
-      {startupProgress && (
-        <div className="fixed inset-0 bg-black/70 backdrop-blur-sm z-[999] flex items-center justify-center">
-          <div className="bg-gradient-to-br from-gray-900/95 to-slate-950/95 backdrop-blur-xl border border-cyan-500/20 p-10 rounded-3xl shadow-2xl w-[800px] max-w-[90vw] max-h-[90vh] overflow-hidden animate-in fade-in zoom-in duration-300 flex flex-col">
-            {!foundGames ? (
-              <div className="flex flex-col items-center space-y-6">
-                {/* Onyx Logo */}
-                <div className="w-24 h-24 animate-pulse">
-                  <svg width="100%" height="100%" viewBox="0 0 512 512" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <defs>
-                      <linearGradient id="onyxGrad" x1="256" y1="20" x2="256" y2="492" gradientUnits="userSpaceOnUse">
-                        <stop offset="0" stopColor="#334155" />
-                        <stop offset="1" stopColor="#020617" />
-                      </linearGradient>
-                      <filter id="glow" x="-50%" y="-50%" width="200%" height="200%">
-                        <feGaussianBlur stdDeviation="8" result="blur" />
-                        <feComposite in="SourceGraphic" in2="blur" operator="over" />
-                      </filter>
-                    </defs>
-
-                    <path d="M256 30 L465 150 V362 L256 482 L47 362 V150 L256 30Z"
-                      fill="url(#onyxGrad)"
-                      stroke="#0ea5e9"
-                      strokeWidth="8"
-                      filter="url(#glow)" />
-
-                    <path d="M256 256 L256 482 M256 256 L47 150 M256 256 L465 150"
-                      stroke="#1e293b"
-                      strokeWidth="4" />
-
-                    <g transform="translate(256, 143) scale(1, 0.58)">
-                      <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                      <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                    </g>
-
-                    <g transform="translate(151, 325) rotate(60) scale(1, 0.58)">
-                      <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                      <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                    </g>
-
-                    <g transform="translate(361, 325) rotate(-60) scale(1, 0.58)">
-                      <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                      <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                    </g>
-
-                    <path d="M256 30 L465 150 L256 256 L47 150 L256 30Z"
-                      fill="white"
-                      fillOpacity="0.1" />
-                  </svg>
-                </div>
-
-                {/* Title */}
-                <div className="text-center">
-                  <h3 className="text-2xl font-bold text-white mb-2">Scanning Game Libraries</h3>
-                  <p className="text-cyan-100/60 text-sm">Checking for new games on startup...</p>
-                </div>
-
-                {/* Progress message - larger and more visible */}
-                <div className="w-full bg-slate-800/50 rounded-xl p-4 border border-cyan-500/10">
-                  <p className="text-cyan-50/90 text-base text-center font-medium break-words">
-                    {startupProgress.message}
-                  </p>
-                </div>
-
-                {/* Progress bar */}
-                <div className="w-full bg-slate-700/50 rounded-full h-2.5 overflow-hidden">
-                  <div className="h-full bg-gradient-to-r from-cyan-500 to-sky-400 rounded-full animate-pulse" style={{ width: '100%' }}></div>
-                </div>
-              </div>
-            ) : (
-              <div className="flex flex-1 min-h-0 flex-col">
-                <div className="flex items-start gap-4 mb-6">
-                  {/* Onyx Logo */}
-                  <div className="w-16 h-16 flex-shrink-0">
-                    <svg width="100%" height="100%" viewBox="0 0 512 512" fill="none" xmlns="http://www.w3.org/2000/svg">
-                      <defs>
-                        <linearGradient id="onyxGrad3" x1="256" y1="20" x2="256" y2="492" gradientUnits="userSpaceOnUse">
-                          <stop offset="0" stopColor="#334155" />
-                          <stop offset="1" stopColor="#020617" />
-                        </linearGradient>
-                        <filter id="glow3" x="-50%" y="-50%" width="200%" height="200%">
-                          <feGaussianBlur stdDeviation="8" result="blur" />
-                          <feComposite in="SourceGraphic" in2="blur" operator="over" />
-                        </filter>
-                      </defs>
-
-                      <path d="M256 30 L465 150 V362 L256 482 L47 362 V150 L256 30Z"
-                        fill="url(#onyxGrad3)"
-                        stroke="#0ea5e9"
-                        strokeWidth="8"
-                        filter="url(#glow3)" />
-
-                      <path d="M256 256 L256 482 M256 256 L47 150 M256 256 L465 150"
-                        stroke="#1e293b"
-                        strokeWidth="4" />
-
-                      <g transform="translate(256, 143) scale(1, 0.58)">
-                        <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                        <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                      </g>
-
-                      <g transform="translate(151, 325) rotate(60) scale(1, 0.58)">
-                        <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                        <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                      </g>
-
-                      <g transform="translate(361, 325) rotate(-60) scale(1, 0.58)">
-                        <circle r="55" stroke="#0ea5e9" strokeWidth="20" strokeOpacity="0.6" fill="none" />
-                        <circle r="55" stroke="#e0f2fe" strokeWidth="8" fill="none" />
-                      </g>
-
-                      <path d="M256 30 L465 150 L256 256 L47 150 L256 30Z"
-                        fill="white"
-                        fillOpacity="0.1" />
-                    </svg>
-                  </div>
-                  <div>
-                    <h2 className="text-2xl font-bold text-white">New Games Found</h2>
-                    <p className="text-gray-400 text-sm mt-1">
-                      The following new games were detected. You can review them in the Importer.
-                    </p>
-                  </div>
-                </div>
-
-                <div className="flex items-center mb-4 px-1">
-                  <span className="text-sm font-medium text-gray-400">
-                    {foundGames.length} {foundGames.length === 1 ? 'game' : 'games'} found
-                  </span>
-                </div>
-
-                <div className="flex-1 min-h-0 overflow-y-auto mb-4 pr-2 -mr-2 space-y-2">
-                  {[...foundGames].sort((a, b) => a.title.localeCompare(b.title)).map((game, index) => (
-                    <div
-                      key={game.id || `game-${index}`}
-                      className="flex items-start gap-4 p-4 bg-gray-800/40 border border-gray-700/30 rounded-xl hover:bg-gray-800/60 transition-colors group"
-                    >
-                      <div className="flex-1 min-w-0">
-                        <div className="flex justify-between items-start gap-2">
-                          <div className="font-semibold text-white truncate">{game.title}</div>
-                          {(game.platform || game.source) && (
-                            <div className="px-2 py-0.5 rounded text-xs font-medium bg-gray-700 text-gray-300 uppercase tracking-wider text-[10px]">
-                              {game.platform || game.source}
-                            </div>
-                          )}
-                        </div>
-                        {(game.exePath || game.installPath) && (
-                          <div className="text-xs text-gray-500 mt-1 font-mono break-all group-hover:text-gray-400 transition-colors">
-                            {game.exePath || game.installPath}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-
-                <div className="flex gap-3 pt-4 border-t border-gray-800 bg-gradient-to-t from-slate-950/90 to-transparent">
-                  <button
-                    onClick={() => {
-                      setFoundGames(null);
-                      setStartupProgress(null);
-                    }}
-                    className="flex-1 px-5 py-2.5 bg-gray-800 hover:bg-gray-700 text-white font-medium rounded-lg transition-colors border border-gray-700"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    onClick={() => {
-                      // Close scanning overlay first for smooth transition
-                      setStartupProgress(null);
-
-                      // Small delay for transition effect
-                      setTimeout(() => {
-                        // Determine app type from the games (use 'other' for mixed sources)
-                        const sources = new Set(foundGames.map((g: any) => g.source));
-                        const appType = sources.size === 1 && sources.has('steam') ? 'steam' :
-                          sources.size === 1 && sources.has('xbox') ? 'xbox' : 'other';
-
-                        setScannedSteamGames(foundGames);
-                        setImportAppType(appType);
-                        setAutoStartScan(true); // Enable auto-scan
-                        setIsImportWorkbenchOpen(true);
-                        setFoundGames(null);
-                      }, 200);
-                    }}
-                    className="flex-1 px-5 py-2.5 bg-gray-700 hover:bg-gray-600 text-white font-medium rounded-lg transition-colors border border-gray-600"
-                  >
-                    Review in Importer
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
+    <PreferencesProvider>
+      <UIProvider>
+        <UpdateProvider>
+          <AppLayout />
+        </UpdateProvider>
+      </UIProvider>
+    </PreferencesProvider>
   );
 }
 
 export default App;
-
-
